% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{anno_gc_vm32}
\alias{anno_gc_vm32}
\title{Mouse Genome Annotation (GC/VM32)}
\format{
A data frame with columns:
\describe{
  \item{id}{Ensembl gene identifier (e.g., "ENSMUSG00000000001").}
  \item{eff_length}{Effective gene length for TPM calculation.}
  \item{gc}{GC content proportion (0-1 range).}
  \item{symbol}{Official gene symbol.}
  \item{mgi_id}{Mouse Genome Informatics identifier.}
  \item{gene_type}{Gene type classification (e.g., "protein_coding", "lncRNA").}
  \item{start}{Genomic start position.}
  \item{end}{Genomic end position.}
  \item{transcript_id}{Transcript identifier (mostly NA in this dataset).}
  \item{ont}{Gene ontology information (mostly NA in this dataset).}
}
}
\source{
Ensembl database for mouse GRCm38/mm10
}
\usage{
data(anno_gc_vm32)
}
\description{
Mouse gene annotation dataset for genome assembly GRCm38/mm10,
containing gene features including GC content and gene length information.
Used for mouse transcriptomic data analysis.
}
\keyword{datasets}
