% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{anno_grch38}
\alias{anno_grch38}
\title{GRCh38 Human Genome Annotation}
\format{
A data frame with 60668 rows and 11 columns:
\describe{
  \item{id}{Ensembl gene identifier.}
  \item{eff_length}{Effective gene length for TPM calculation.}
  \item{gc}{GC content proportion.}
  \item{entrez}{Entrez Gene ID.}
  \item{symbol}{Official gene symbol.}
  \item{chr}{Chromosome name.}
  \item{start}{Genomic start position.}
  \item{end}{Genomic end position.}
  \item{strand}{Genomic strand.}
  \item{biotype}{Gene biotype classification.}
  \item{description}{Gene description.}
}
}
\source{
Ensembl release 104 (GRCh38.p13)
}
\usage{
data(anno_grch38)
}
\description{
Gene annotation for human GRCh38/hg38 genome assembly, used for expression
normalization and gene identifier mapping in IOBR functions.
}
\keyword{datasets}
