% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpcounter.R
\name{appendSignatures}
\alias{appendSignatures}
\title{Append Signatures to Expression Matrix}
\usage{
appendSignatures(xp, markers)
}
\arguments{
\item{xp}{An expression matrix with features in rows and samples in columns.}

\item{markers}{A list whose names are cellular populations' names and elements are character vectors of features.}
}
\value{
Matrix with the summarized expression of each marker feature set in rows.
}
\description{
Takes as input an expression matrix and a list of marker features and returns summarized expression values.
}
\author{
Etienne Becht
}
