% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPCA.R
\name{batchPCA}
\alias{batchPCA}
\title{batch PCA}
\usage{
batchPCA(
  indata,
  batch,
  pos = "bottomright",
  xy = c(1, 2),
  cols = NULL,
  showID = FALSE,
  cex = 1,
  showLegend = T,
  main.title = NULL
)
}
\arguments{
\item{indata}{input data, expression set data}

\item{batch}{a string equal to length of indata}

\item{pos}{legend position}

\item{xy}{dimension of PCA}

\item{cols}{colors of batch}

\item{showID}{default is false}

\item{cex}{cex of each point}

\item{showLegend}{default is true}

\item{main.title}{title of plot}
}
\description{
batch PCA
}
\author{
Xiaofan Lu, Dongqiang Zeng
}
