% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_group.R
\name{calculate_break_month}
\alias{calculate_break_month}
\title{Break Time Into Blocks}
\usage{
calculate_break_month(input, block = 6, time_type = c("month", "day"))
}
\arguments{
\item{input}{Numeric vector of time durations.}

\item{block}{Number of blocks. Default is 6.}

\item{time_type}{Units: "month" or "day". Default is "month".}
}
\value{
Numeric vector of breakpoints, rounded to nearest multiple of 5.
}
\description{
Divides time duration into specified blocks for analysis.
}
\examples{
time_data <- c(24, 36, 12, 48)
blocks <- calculate_break_month(input = time_data)
}
\author{
Dongqiang Zeng
}
