% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseq_helper.R
\name{DClsei}
\alias{DClsei}
\title{Constrained Least Squares Deconvolution}
\usage{
DClsei(b, A, G, H, scaling)
}
\arguments{
\item{b}{Observation vector.}

\item{A}{Signature matrix.}

\item{G}{Constraint matrix.}

\item{H}{Constraint vector.}

\item{scaling}{Scaling factor.}
}
\value{
Estimated cell fractions.
}
\description{
Constrained Least Squares Deconvolution
}
\keyword{internal}
