% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseq_helper.R
\name{DCrr}
\alias{DCrr}
\title{Robust Regression Deconvolution}
\usage{
DCrr(b, A, method, scaling)
}
\arguments{
\item{b}{Observation vector.}

\item{A}{Signature matrix.}

\item{method}{Robust regression method.}

\item{scaling}{Scaling factor.}
}
\value{
Estimated cell fractions.
}
\description{
Robust Regression Deconvolution
}
\keyword{internal}
