% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_constru_sig.R
\name{deconvo_constru_sig}
\alias{deconvo_constru_sig}
\title{Estimate the fraction of cell types using defined reference genes}
\usage{
deconvo_constru_sig(
  eset,
  reference,
  scale_reference = TRUE,
  method = "svm",
  arrays = TRUE,
  perm = 500
)
}
\arguments{
\item{eset}{expression data with matched gene id of reference}

\item{reference}{immune cell gene matrix; eg lm22, lm6 or can be generate using generateRef/generateRef_rnaseq}

\item{scale_reference}{a logical value indicating whether the reference be scaled or not. If TRUE, the value in reference file will be centered and scaled in row direction.}

\item{method}{method must be set to either 'svm' or 'lsei'}

\item{arrays}{a logical value. If TRUE, the columns of the input data will be normalized to have the same quantiles.}

\item{perm}{Set permutations for statistical analysis (≥100 permutations recommended).}
}
\value{

}
\description{
Estimate the fraction of cell types using defined reference genes
}
\examples{

data("eset_crc")
data("lm22")
## eset_crc has already been normalized, so here normalize = FALSE
til_svm <- deconvo_constru_sig(eset = eset_crc[, 1:50], reference = lm22, scale_reference = T, method = "svm", arrays = F)
til_lsei <- deconvo_constru_sig(eset= eset_crc[, 1:50], reference = lm22_ref, scale_reference = T, method = "lsei",arrays = F)
}
