% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_mcpcounter}
\alias{deconvo_mcpcounter}
\title{Deconvolve Immune Microenvironment Using MCP-Counter}
\usage{
deconvo_mcpcounter(eset, project = NULL)
}
\arguments{
\item{eset}{Gene expression matrix with HGNC symbols as row names.}

\item{project}{Optional project name. Default is `NULL`.}
}
\value{
Data frame with MCP-counter scores. Columns suffixed with
  `_MCPcounter`.
}
\description{
Estimates immune cell abundances using MCP-counter.
}
\examples{
eset_stad <- load_data("eset_stad")
anno_grch38 <- load_data("anno_grch38")
eset <- anno_eset(eset = eset_stad, annotation = anno_grch38, probe = "id")
eset <- eset[1:500, 1:3]
mcp_result <- deconvo_mcpcounter(eset = eset, project = "TCGA-STAD")
head(mcp_result)
}
\author{
Dongqiang Zeng
}
