% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{deg}
\alias{deg}
\title{Single-cell RNA-seq Differential Expression Analysis Results}
\format{
A data frame with 3,212 rows (genes) and 7 columns (statistics):
\describe{
  \item{p_val}{Raw p-value for differential expression test}
  \item{avg_log2FC}{Average log2 fold-change of gene expression}
  \item{pct.1}{Percentage of cells expressing the gene in cluster 1}
  \item{pct.2}{Percentage of cells expressing the gene in cluster 2}
  \item{p_val_adj}{Adjusted p-value (e.g., Bonferroni, BH FDR correction)}
  \item{cluster}{Cell cluster or cell type identifier (e.g., "Epithelial cells 2")}
  \item{gene}{Gene symbol (e.g., "IGFBP3", "PCDH7")}
}
}
\usage{
data(deg)
}
\description{
Example dataset containing differential expression analysis results
from single-cell RNA sequencing (scRNA-seq) analysis.
This dataset serves as input for signature analysis functions in the IOBR package,
particularly \code{\link{get_sig_sc}} and \code{\link{sig_gsea}}.
}
\keyword{datasets}
