% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpcounter.R
\name{.appendSignatures}
\alias{.appendSignatures}
\title{Append Signatures to Expression Matrix}
\usage{
.appendSignatures(xp, markers)
}
\arguments{
\item{xp}{Expression matrix with features in rows and samples in columns.}

\item{markers}{List of marker gene vectors for each cell population.}
}
\value{
Matrix with summarized expression values.
}
\description{
Calculates mean expression for each marker feature set.
}
\keyword{internal}
