% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mut_matrix.R
\name{.build_mut_matrices}
\alias{.build_mut_matrices}
\title{Build Mutation Matrices from MAF Data}
\usage{
.build_mut_matrices(mut, category)
}
\arguments{
\item{mut}{Data frame with mutation data.}

\item{category}{Character. Mutation category.}
}
\value{
List of mutation matrices.
}
\description{
Build Mutation Matrices from MAF Data
}
\keyword{internal}
