% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mut_matrix.R
\name{.create_mut_matrix}
\alias{.create_mut_matrix}
\title{Create Single Mutation Matrix}
\usage{
.create_mut_matrix(mut)
}
\arguments{
\item{mut}{Data frame with mutation data.}
}
\value{
Mutation matrix.
}
\description{
Create Single Mutation Matrix
}
\keyword{internal}
