% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_barplot.R
\name{enrichment_barplot}
\alias{enrichment_barplot}
\title{Enrichment bar plot with two directions}
\usage{
enrichment_barplot(
  up_terms,
  down_terms,
  terms = "Description",
  pvalue = "pvalue",
  group = "group",
  palette = "jama",
  title = "Gene Ontology Enrichment",
  width_wrap = 30,
  font_terms = 15
)
}
\arguments{
\item{up_terms}{A data frame containing the information for up-regulated terms}

\item{down_terms}{A data frame containing the information for down-regulated terms or genes.}

\item{terms}{The column name in the data frames that specifies the term or gene description. By default, it is set to "Description".}

\item{pvalue}{The column name in the data frames that specifies the p-value. By default, it is set to "pvalue".}

\item{group}{The column name in the data frames that specifies the group or condition. By default, it is set to "group". If this column is not present in the data frames, the function will automatically assign a group based on the up_terms and down_terms.}

\item{palette}{The color palette to use for the barplot. By default, it is set to "jama".}

\item{title}{The title for the barplot. By default, it is set to "Gene Ontology Enrichment".}

\item{width_wrap}{The maximum width for wrapping the pathway names on the x-axis labels. By default, it is set to 30 characters.}

\item{font_terms}{The font size for the x-axis and y-axis labels. By default, it is set to 15.}
}
\description{
Enrichment bar plot with two directions
}
\author{
Dongqiang Zeng
}
