% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_helper.R
\name{filterCommonGenes}
\alias{filterCommonGenes}
\title{filterCommonGenes}
\usage{
filterCommonGenes(input.f, output.f, id = c("GeneSymbol", "EntrezID"))
}
\arguments{
\item{input.f}{input.f}

\item{output.f}{output.f}

\item{id}{"GeneSymbol", "EntrezID"}
}
\description{
filterCommonGenes
}
