% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mutations.R
\name{find_mutations}
\alias{find_mutations}
\title{Title}
\usage{
find_mutations(
  mutation_matrix,
  signature_matrix,
  id_signature_matrix = "ID",
  signature,
  mutation_freq = 0.05,
  plot = TRUE,
  method = "multi",
  save_path = NULL,
  palette = "paired3"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Title
}
\examples{
choose top 10 genes
willcoxon test
choose top 10 genes
willcoxon test
choose top 10 genes
}
