% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseq_helper.R
\name{fixMixture}
\alias{fixMixture}
\title{Fix Expression Mixture Matrix}
\usage{
fixMixture(mix.mat, arrays = FALSE)
}
\arguments{
\item{mix.mat}{Expression matrix with genes as rows.}

\item{arrays}{Logical indicating if data is from arrays. Default is FALSE.}
}
\value{
Processed expression matrix.
}
\description{
Processes expression matrix by mapping genes, converting log values,
quantile normalization (if arrays), and TPM normalization.
}
\keyword{internal}
