% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_default_mirrors}
\alias{get_default_mirrors}
\title{Get Default Download Mirrors}
\usage{
get_default_mirrors()
}
\value{
Character vector of mirror URLs.
}
\description{
Returns the default list of download mirrors.
}
\keyword{internal}
