% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobr_deconvo_pieline.R
\name{iobr_deconvo_pieline}
\alias{iobr_deconvo_pieline}
\title{Title iobr_deconvo_pieline}
\usage{
iobr_deconvo_pieline(
  eset,
  project,
  array,
  tumor_type,
  path = "1-TME",
  permutation = 1000
)
}
\arguments{
\item{eset}{}

\item{project}{}

\item{array}{}

\item{tumor_type}{}

\item{path}{}

\item{permutation}{}
}
\description{
Title iobr_deconvo_pieline
}
