% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_helper.R
\name{ipsmap}
\alias{ipsmap}
\title{Calculate Immunophenoscore}
\usage{
ipsmap(x)
}
\arguments{
\item{x}{A numeric value representing the input score.}
}
\value{
A numeric value between 0 and 10.
}
\description{
Maps input score to Immunophenoscore (IPS) on a 0-10 scale.
}
\examples{
ips <- ipsmap(2.5)
ips <- ipsmap(-1)
ips <- ipsmap(5)
}
