% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{list_github_datasets}
\alias{list_github_datasets}
\title{List Available GitHub Datasets}
\usage{
list_github_datasets()
}
\value{
Character vector of available dataset names.
}
\description{
List Available GitHub Datasets
}
\examples{
list_github_datasets()
}
