% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{list_iobr_mirrors}
\alias{list_iobr_mirrors}
\title{List Current Download Mirrors}
\usage{
list_iobr_mirrors()
}
\value{
Character vector of mirror URLs.
}
\description{
Returns the current list of download mirrors.
}
\examples{
list_iobr_mirrors()
}
