% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{load_data}
\alias{load_data}
\title{Load IOBR Datasets}
\usage{
load_data(name)
}
\arguments{
\item{name}{A dataset name.}
}
\value{
Dataset, typically `list` or `data.frame`.
}
\description{
Load IOBR Datasets
}
\examples{
load_data("deg")
}
