% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mut_matrix.R
\name{make_mut_matrix}
\alias{make_mut_matrix}
\title{Title}
\usage{
make_mut_matrix(
  maf = NULL,
  maf_data = NULL,
  isTCGA = TRUE,
  category = "multi",
  Tumor_Sample_Barcode = "Tumor_Sample_Barcode",
  Hugo_Symbol = "Hugo_Symbol",
  Variant_Classification = "Variant_Classification",
  Variant_Type = "Variant_Type"
)
}
\arguments{
\item{maf}{}

\item{maf_data}{}

\item{isTCGA}{}

\item{category}{}

\item{Tumor_Sample_Barcode}{}

\item{Hugo_Symbol}{}

\item{Variant_Classification}{}

\item{Variant_Type}{}
}
\value{

}
\description{
Title
}
