% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseq_helper.R
\name{makeQN}
\alias{makeQN}
\title{Quantile Normalization}
\usage{
makeQN(mix.mat)
}
\arguments{
\item{mix.mat}{Expression matrix.}
}
\value{
Quantile normalized matrix.
}
\description{
Quantile Normalization
}
\keyword{internal}
