% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_helper.R
\name{mapcolors}
\alias{mapcolors}
\title{Map Score to Color}
\usage{
mapcolors(x, my_palette = NULL)
}
\arguments{
\item{x}{A numeric value to be mapped to a color.}

\item{my_palette}{Color palette vector (should have 1001 colors).
Default uses red-white-blue gradient.}
}
\value{
A color from the palette.
}
\description{
Maps a numeric input value to a color from a predefined palette.
}
\examples{
\donttest{
# Create palette with 1001 colors
my_palette <- grDevices::colorRampPalette(c("blue", "white", "red"))(1001)
color <- mapcolors(2, my_palette)
color <- mapcolors(-2, my_palette)
}
}
