% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseq_helper.R
\name{mapGenes}
\alias{mapGenes}
\title{Map Gene Names to Approved Symbols}
\usage{
mapGenes(mydata)
}
\arguments{
\item{mydata}{Expression matrix with gene symbols as rownames.}
}
\value{
Matrix with mapped gene symbols.
}
\description{
Maps gene symbols to approved HGNC symbols, handling withdrawn and
previous symbols.
}
\keyword{internal}
