% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_helper.R
\name{merge_duplicates}
\alias{merge_duplicates}
\title{Merging the duplicates from the input matrix.}
\usage{
merge_duplicates(mat, warn = TRUE, in_type = NULL)
}
\arguments{
\item{mat}{matrix}

\item{warn}{warning}

\item{in_type}{default is null}
}
\description{
In case there are some duplicate rownames in the input matrix (\emph{mat}),
this function will return a similar matrix with unique rownames and the
duplicate cases will be merged together based on the median values. When
\emph{warn} is true a warning message will be written in case there are
duplicates, this warning message will include the \emph{in_type} string to
indicate the current type of the input matrix.
}
\keyword{internal}
