% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{microenvironmentScores}
\alias{microenvironmentScores}
\title{Calculate Microenvironment Scores}
\usage{
microenvironmentScores(adjustedScores)
}
\arguments{
\item{adjustedScores}{Adjusted xCell scores matrix.}
}
\value{
Matrix with additional microenvironment score rows.
}
\description{
Calculates combined immune and stroma scores from adjusted xCell scores.
}
\keyword{internal}
