% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{null_models}
\alias{null_models}
\title{NULL Model Coefficients for MCPcounter}
\format{
A `data.frame` with cell types in rows and coefficients in columns.
}
\usage{
data(null_models)
}
\description{
NULL Model Coefficients for MCPcounter
}
\keyword{datasets}
