% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palettes}
\alias{palettes}
\title{Choose colors.}
\usage{
palettes(
  category = "box",
  palette = "nrc",
  alpha = 1,
  counts = 50,
  show_col = TRUE,
  show_message = FALSE
)
}
\arguments{
\item{category}{choose what type of plot you want to color,There are seven categories you can choose: `box`, `continue2`, `continue`, `random`, `heatmap`, `heatmap3`, `tidyheatmap`}

\item{palette}{Each category has some options to choose}

\item{alpha}{intensity of the color}

\item{counts}{number of colors, only affect continue colors}

\item{show_col}{logical variable, if TRUE, color will be print and show in the R studio}

\item{show_message}{print palettes that can be choose}
}
\value{

}
\description{
Choose colors.
}
\author{
Dongqiang Zeng
}
