% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_names.R
\docType{data}
\name{patterns_to_na}
\alias{patterns_to_na}
\title{Default Pattern List for Name Cleaning}
\format{
A character vector of length 12.
}
\usage{
patterns_to_na
}
\value{
Character vector of patterns to remove.
}
\description{
A character vector of common substrings to remove from feature names.
Used in [remove_names()] and other helper functions.
}
\examples{
# View default patterns
patterns_to_na
}
\keyword{datasets}
