% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModel.R
\name{PlotAUC}
\alias{PlotAUC}
\title{Plot AUC ROC Curves}
\usage{
PlotAUC(
  train.x,
  train.y,
  test.x,
  test.y,
  model,
  modelname,
  cols = NULL,
  palette = "jama"
)
}
\arguments{
\item{train.x}{Training predictors.}

\item{train.y}{Training outcomes.}

\item{test.x}{Testing predictors.}

\item{test.y}{Testing outcomes.}

\item{model}{Fitted model.}

\item{modelname}{Model name for title.}

\item{cols}{Optional color vector.}

\item{palette}{Color palette name.}
}
\value{
ggplot object of ROC curve.
}
\description{
Generates ROC curves for model evaluation.
}
\examples{
\dontrun{
PlotAUC(train.x, train.y, test.x, test.y, fitted_model, "MyModel")
}
}
