% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_helper.R
\name{plotPurity}
\alias{plotPurity}
\title{plotPurity}
\usage{
plotPurity(
  scores,
  samples = "all_samples",
  platform = c("affymetrix", "agilent", "illumina"),
  output.dir = "estimated_purity_plots"
)
}
\arguments{
\item{scores}{A character string specifying the path to the input file containing ESTIMATE scores. The file should be a tab-separated table with appropriate headers.}

\item{samples}{A character vector specifying the sample names to plot. The default is "all_samples", which plots all samples in the input file.}

\item{platform}{A character string specifying the platform used for data collection. Can be "affymetrix", "agilent", or "illumina". Currently, only "affymetrix" is implemented.}

\item{output.dir}{A character string specifying the directory to save the output plots. The default is "estimated_purity_plots".}
}
\value{
No return value. The function generates and saves scatterplots in the specified output directory.
}
\description{
This function generates scatterplots of tumor purity based on ESTIMATE
scores for given samples.
}
\examples{
# Create a sample ESTIMATE score matrix
scores_data <- data.frame(
  Sample1 = c(100, 200, 500, 0.80),
  Sample2 = c(120, 220, 450, 0.70),
  Sample3 = c(140, 240, 600, 0.90),
  row.names = c(
    "StromalScore", "ImmuneScore", "ESTIMATEScore",
    "TumorPurity"
  ),
  check.names = FALSE
)

# Write to a temporary GCT file
scores_file <- tempfile(fileext = ".gct")
outputGCT(scores_data, scores_file)
}
