% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{PlotTimeROC}
\alias{PlotTimeROC}
\title{Plot Time-Dependent ROC Curves}
\usage{
PlotTimeROC(
  train.x,
  train.y,
  test.x,
  test.y,
  model,
  modelname,
  cols = NULL,
  palette = "jama"
)
}
\arguments{
\item{train.x}{Matrix or data frame of training predictors.}

\item{train.y}{Training survival outcomes (time and status).}

\item{test.x}{Matrix or data frame of testing predictors.}

\item{test.y}{Testing survival outcomes (time and status).}

\item{model}{Fitted survival model object.}

\item{modelname}{Character string for model identification.}

\item{cols}{Optional vector of colors for plotting.}

\item{palette}{Character string specifying color palette. Default is `"jama"`.}
}
\value{
A `ggplot` object representing the ROC curve plot.
}
\description{
Generates time-dependent ROC curves for evaluating prognostic accuracy of survival models.
Plots training and testing ROC curves at the 90th percentile survival time.
}
\examples{
\dontrun{
PlotTimeROC(train.x, train.y, test.x, test.y, fitted_model, "Cox Model")
}
}
\author{
Dongqiang Zeng
}
