% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseq_helper.R
\name{quanTIseq}
\alias{quanTIseq}
\title{Run quanTIseq Deconvolution}
\usage{
quanTIseq(currsig, currmix, scaling, method)
}
\arguments{
\item{currsig}{Signature matrix.}

\item{currmix}{Mixture matrix.}

\item{scaling}{Scaling factors.}

\item{method}{Deconvolution method: "lsei", "hampel", "huber", or "bisquare".}
}
\value{
Deconvolution results matrix.
}
\description{
Run quanTIseq Deconvolution
}
\keyword{internal}
