% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiseq_helper.R
\name{quantiseq_helper}
\alias{quantiseq_helper}
\title{Helper Functions for quanTIseq}
\description{
This package provides helper functions for quanTIseq deconvolution method.
Source code adapted from https://github.com/FFinotello/quanTIseq.
}
