% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{rawEnrichmentAnalysis}
\alias{rawEnrichmentAnalysis}
\title{Calculate Raw xCell Enrichment Scores}
\usage{
rawEnrichmentAnalysis(expr, signatures, genes, file.name = NULL)
}
\arguments{
\item{expr}{Gene expression data matrix with row names as gene symbols.}

\item{signatures}{GMT object of signatures.}

\item{genes}{Character vector of genes to use.}

\item{file.name}{Character string for saving scores. Default is `NULL`.}
}
\value{
Matrix of raw xCell scores.
}
\description{
Returns the raw xCell cell types enrichment scores using ssGSEA.
}
\keyword{internal}
