% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_iobr.R
\name{rbind_iobr}
\alias{rbind_iobr}
\title{rbind_iobr}
\usage{
rbind_iobr(data1, data2, data3 = NULL)
}
\arguments{
\item{data1}{The first dataset to be combined. It should be a data.frame or a matrix.}

\item{data2}{The second dataset to be combined. It should be a data.frame or a matrix.}

\item{data3}{(Optional) An additional dataset to be combined. It should be a data.frame or a matrix. Default value is NULL.}
}
\value{
The combined dataset resulting from the 'rbind' operation on the input datasets.
}
\description{
This function combines multiple datasets vertically using the 'rbind' operation. It takes in two required parameters, 'data1' and 'data2', and an optional parameter, 'data3'.
}
