% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_signatures.R
\name{read_signatures}
\alias{read_signatures}
\title{Reading signature gene matrix as list}
\usage{
read_signatures(file_name, save_signature = FALSE, output_name = "signatures")
}
\arguments{
\item{file_name}{the file name of csv file with signature name as column name and genes in row}

\item{save_signature}{logical, save siganture list as RData}

\item{output_name}{string of name to save signature RData}
}
\value{
list of signatures
}
\description{
Reading signature gene matrix as list
}
\examples{
mysignature<-read_signatures(file_name = "signatures.csv",save_signature = TRUE,output_name = "mysignatures")
reading each column and transfer it to a list
standerdized the name of list
}
\author{
Dongqiang Zeng
}
