% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{reset_iobr_cache_dir}
\alias{reset_iobr_cache_dir}
\title{Reset IOBR Cache Directory to Default}
\usage{
reset_iobr_cache_dir()
}
\value{
Invisibly returns the default cache directory path.
}
\description{
Resets the cache directory to the default system location.
}
\examples{
reset_iobr_cache_dir()
}
