% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{reset_iobr_mirrors}
\alias{reset_iobr_mirrors}
\title{Reset Download Mirrors to Default}
\usage{
reset_iobr_mirrors()
}
\value{
Invisibly returns the default mirror list.
}
\description{
Resets the download mirror list to the default values.
}
\examples{
reset_iobr_mirrors()
}
