% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_helper.R
\name{scaleCounts}
\alias{scaleCounts}
\title{Scaling raw counts from each sample.}
\usage{
scaleCounts(counts, sigGenes = NULL, renormGenes = NULL, normFact = NULL)
}
\description{
Normalizing the sum of counts from each sample to 1e6.
}
\details{
Function taking a matrix (\emph{genes} x \emph{samples}) of counts as input
and returning the scaled counts for the subset signature genes (or all genes
if it sigGenes is \code{NULL}), with the scaled counts computed based on all
the 'renormGenes' (or all genes if it is NULL). The renormalization is made
independently for each sample, to have the sum of each columns over the
renormGenes equal to 1e6.

normFact, if not null, is used as the normalization factor instead of
 the colSums (used to renormalize the refProfiles.var by the same amount
 than the refProfiles).
}
\keyword{internal}
