% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score.R
\docType{data}
\name{signature_score_calculation_methods}
\alias{signature_score_calculation_methods}
\title{Signature Score Calculation Methods}
\format{
Named character vector:
\describe{
  \item{PCA}{Principal Component Analysis method ("pca")}
  \item{ssGSEA}{Single-sample Gene Set Enrichment Analysis ("ssgsea")}
  \item{z-score}{Z-score transformation method ("zscore")}
  \item{Integration}{Integration of multiple methods ("integration")}
}
}
\usage{
signature_score_calculation_methods
}
\description{
A named vector of supported methods for calculating signature scores.
}
\keyword{datasets}
