% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{spillOver}
\alias{spillOver}
\title{Adjust Scores Using Spillover Compensation}
\usage{
spillOver(transformedScores, K, alpha = 0.5, file.name = NULL)
}
\arguments{
\item{transformedScores}{Transformed scores from transformScores.}

\item{K}{Spillover matrix.}

\item{alpha}{Spillover alpha parameter. Default is `0.5`.}

\item{file.name}{Character string for saving scores. Default is `NULL`.}
}
\value{
Adjusted xCell scores matrix.
}
\description{
Adjusts xCell scores using spillover compensation matrix.
}
\keyword{internal}
