% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer_info}
\alias{timer_info}
\title{Source code for the TIMER deconvolution method.}
\usage{
timer_info(string)
}
\arguments{
\item{string}{Character. Message to be displayed.}
}
\value{
None; used for its side effect of printing a message.
}
\description{
Formats and displays informational messages for timing or logging purposes.
Useful for tracking progress or stages of execution within scripts.
}
\details{
This code is adapted from https://github.com/hanfeisun/TIMER, which
again is an adapted version of the original TIMER source code
from http://cistrome.org/TIMER/download.html.

The method is described in Li et al. Genome Biology 2016;17(1):174., PMID
27549193.
Display Timer Information Messages
}
\examples{
timer_info("Data processing started.")
}
\author{
Bo Li
}
