% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{TimerINFO}
\alias{TimerINFO}
\title{Display Timer Information Messages (Deprecated)}
\usage{
TimerINFO(string)
}
\arguments{
\item{string}{Character. Message to be displayed.}
}
\value{
None; used for its side effect of printing a message.
}
\description{
This function is deprecated. Use [timer_info()] instead.
}
\examples{
TimerINFO("Data processing started.")
}
\author{
Bo Li
}
