% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform features with Inf or NA to other number}
\usage{
transform_data(data, feature, data_type = c("NA", "Inf", "zero"), into = 0)
}
\arguments{
\item{data}{A data frame containing the data to be transformed.}

\item{feature}{A vector of column names in 'data' where transformations will be applied.}

\item{data_type}{The type of data to be transformed: "NA", "Inf", or "zero".}

\item{into}{The value to replace 'data_type' with; default is 0. If 'into' is "mean", it replaces with the column mean.}
}
\value{
Returns the data frame with specified transformations applied to the indicated features.
}
\description{
Transform features with Inf or NA to other number
}
\examples{
data_matrix <- data.frame(A = c(1, 2, NA, 4, Inf), B = c(Inf, 2, 3, 4, 5), C = c(0, 0, 0, 1, 2))
# Replace NAs with 0
transform_data(data_matrix, feature = c("A", "B"), data_type = "NA")
# Replace Infinite values with the mean of the column
transform_data(data_matrix, feature = c("A", "B"), data_type = "Inf", into = "mean")
# Replace zeros with -1 in column C
transform_data(data_matrix, feature = "C", data_type = "zero", into = -1)
}
\author{
Dongqiang Zeng
}
