% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{transformScores}
\alias{transformScores}
\title{Transform Raw Scores to Fractions}
\usage{
transformScores(scores, fit.vals, scale = TRUE, fn = NULL)
}
\arguments{
\item{scores}{Raw scores from rawEnrichmentAnalysis.}

\item{fit.vals}{Calibration values from spill object.}

\item{scale}{Logical indicating whether to use scaling. Default is `TRUE`.}

\item{fn}{Character string for saving scores. Default is `NULL`.}
}
\value{
Transformed xCell scores matrix.
}
\description{
Transforms raw xCell scores to estimated cell fractions.
}
\keyword{internal}
