% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{xCellSignifcanceBetaDist}
\alias{xCellSignifcanceBetaDist}
\title{Calculate Significance P-values Using Beta Distribution}
\usage{
xCellSignifcanceBetaDist(
  scores,
  beta_params = NULL,
  rnaseq = TRUE,
  file.name = NULL
)
}
\arguments{
\item{scores}{xCell scores matrix.}

\item{beta_params}{Pre-calculated beta distribution parameters.}

\item{rnaseq}{Logical for RNA-seq vs array parameters. Default is `TRUE`.}

\item{file.name}{Character string for saving p-values. Default is `NULL`.}
}
\value{
Matrix of p-values.
}
\description{
Calculates FDR-adjusted p-values for the null hypothesis that a cell type
is not present in the mixture.
}
\keyword{internal}
