% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{xCellSignifcanceRandomMatrix}
\alias{xCellSignifcanceRandomMatrix}
\title{Calculate Significance Using Random Matrix}
\usage{
xCellSignifcanceRandomMatrix(
  scores,
  expr,
  spill,
  alpha = 0.5,
  nperm = 250,
  file.name = NULL
)
}
\arguments{
\item{scores}{xCell scores matrix.}

\item{expr}{Input expression matrix.}

\item{spill}{Spillover object.}

\item{alpha}{Spillover alpha parameter. Default is `0.5`.}

\item{nperm}{Number of permutations. Default is `250`.}

\item{file.name}{Character string for saving p-values. Default is `NULL`.}
}
\value{
List containing p-values, shuffled xCell scores, shuffled expression,
  and beta distributions.
}
\description{
Calculates FDR-adjusted p-values using a random shuffled matrix.
}
\keyword{internal}
