% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_riskscore.R
\name{add_riskscore}
\alias{add_riskscore}
\title{Add Risk Score to Dataset}
\usage{
add_riskscore(
  input,
  family = "cox",
  target = NULL,
  time = NULL,
  status = NULL,
  vars,
  new_var_name = "riskscore"
)
}
\arguments{
\item{input}{A data frame containing the data to be analyzed.}

\item{family}{A character string specifying whether to use "cox" for Cox regression
or "binary" for logistic regression. Defaults to "cox".}

\item{target}{The name of the target variable (only used if family = "binary").}

\item{time}{The name of the time to event variable (only used if family = "cox").}

\item{status}{The name of the event status variable (only used if family = "cox").}

\item{vars}{A vector of variable names to include in the model.}

\item{new_var_name}{The name of the new variable to be created for storing risk scores.
Defaults to "riskscore".}
}
\value{
The input data frame with an additional column containing the risk scores.
}
\description{
This function computes a risk score for each observation in the dataset
based on the provided Cox proportional hazards model or logistic regression.
}
\examples{
data(ovarian)
ovarian$rscore <- add_riskscore(ovarian, time = "time", status = "status", vars = c("resid.ds", "rx", "ecog.ps"))
}
\author{
Dongqiang Zeng
}
