% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_eset.R
\name{anno_eset}
\alias{anno_eset}
\title{Annotating gene expression matrix and remove duplicated genes}
\usage{
anno_eset(
  eset,
  annotation,
  symbol = "symbol",
  probe = "probe_id",
  method = "mean"
)
}
\arguments{
\item{eset}{gene expression matrix with probe or ensembl, entrez id in the rownames}

\item{annotation}{annotation file with gene symbol and other ids matched to rowname of eset
user can choose anno_hug133plus2, anno_rnaseq and anno_illumina as input}

\item{symbol}{column name of gene symbol in annotation file}

\item{probe}{column name of relevant id in annotation file}

\item{method}{method used to remove duplicated genes}
}
\value{
modified gene expression set
}
\description{
Annotating gene expression matrix and remove duplicated genes
}
