% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assimilate_data.R
\name{assimilate_data}
\alias{assimilate_data}
\title{Assimilate data with similar columns}
\usage{
assimilate_data(data_a, data_b)
}
\arguments{
\item{data_a}{original data frame}

\item{data_b}{manipulated data frame}
}
\value{
data_b
}
\description{
Assimilate data with similar columns
}
\examples{
pdata_a <- as.data.frame(array(3, c(5, 9)))
colnames(pdata_a) <- LETTERS[1:dim(pdata_a)[2]]
pdata_b <- as.data.frame(array(2, c(7, 4)))
colnames(pdata_b) <- c("A", "C", "E", "F")
assimilate_data(data_a = pdata_a, data_b = pdata_b)

}
