% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assimilate_data.R
\name{assimilate_data}
\alias{assimilate_data}
\title{Assimilate data with similar columns}
\usage{
assimilate_data(data_a, data_b)
}
\arguments{
\item{data_a}{originial dataframe}

\item{data_b}{manipulat dataframe}
}
\value{
data_b
}
\description{
Assimilate data with similar columns
}
\examples{

pdata_a<-as.data.frame(array(NA,c(5,9)));pdata_a
colnames(pdata_a)<-LETTERS[1:dim(pdata_a)[2]];pdata_a
pdata_b<-as.data.frame(array(NA,c(7,4)));pdata_b
colnames(pdata_b)<-c("A","C","E","F");pdata_b
assimilate_data(data_a = pdata_a,data_b = pdata_b)

}
