% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assimilate_data.R
\name{assimilate_data}
\alias{assimilate_data}
\title{Harmonize Two Data Frames by Column Structure}
\usage{
assimilate_data(data_a, data_b)
}
\arguments{
\item{data_a}{Data frame. Reference data frame whose column structure should be matched.}

\item{data_b}{Data frame. Data frame to be conformed to `data_a`.}
}
\value{
Data frame `data_b` with added missing columns (NA-filled) and reordered to match `data_a`.
}
\description{
Adds missing columns (filled with NA) to a secondary data frame so that its column set and order match a reference data frame.
}
\examples{
pdata_a <- as.data.frame(array(3, c(5, 9)))
colnames(pdata_a) <- LETTERS[1:dim(pdata_a)[2]]
pdata_b <- as.data.frame(array(2, c(7, 4)))
colnames(pdata_b) <- c("A", "C", "E", "F")
assimilate_data(data_a = pdata_a, data_b = pdata_b)

}
