% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_cor.R
\name{batch_cor}
\alias{batch_cor}
\title{Batch to conduct correlation analysis with multiple features}
\usage{
batch_cor(data, target, feature, method = "spearman")
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{target}{A character specifying the name of the target variable.}

\item{feature}{A character vector specifying the names of the feature variables.}

\item{method}{A character specifying the correlation method to be used. Default value is "spearman".}
}
\value{
A tibble containing the feature names, p-values, correlation coefficients, adjusted p-values, log10 p-values, and significance stars.
}
\description{
This function performs batch correlation analysis between a target variable and multiple feature variables in a dataset.
}
\examples{

# Loading TCGA-STAD microenvironment signature data
data("sig_stad", package = "IOBR")
# Finding microenvironmental scores associated with CD8 T cells
batch_cor(data = sig_stad, target = "CD_8_T_effector", feature = colnames(sig_stad)[69:ncol(sig_stad)])
}
