% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_cor.R
\name{batch_cor}
\alias{batch_cor}
\title{Batch to conduct correlation analysis with multiple features}
\usage{
batch_cor(data, target, feature, method = "spearman")
}
\arguments{
\item{data}{signature matrix with multiple features}

\item{target}{name of group}

\item{feature}{feature used to comparison}

\item{method}{method must be either `spearman` or `pearson`}
}
\value{

}
\description{
Batch to conduct correlation analysis with multiple features
}
