% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_kruskal.R
\name{batch_kruskal}
\alias{batch_kruskal}
\title{Title}
\usage{
batch_kruskal(data, group, feature = NULL, feature_manipulation = FALSE)
}
\arguments{
\item{feature_manipulation}{}
}
\description{
Title
}
