% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_pcc.R
\name{batch_pcc}
\alias{batch_pcc}
\title{Batch way to calculate the partial correlation coefficient}
\usage{
batch_pcc(data, target, feature, method = "pearson", interference)
}
\arguments{
\item{data}{matrix;data signature matrix with multiple features}

\item{target}{a character vector; target name of group}

\item{feature}{character vectors; feature used to comparison}

\item{method}{vector; one of "pearson"(default), "spearman" or "kendall"}

\item{interference}{a numeric vector, matrix or data frame; variabes used to control when calculate the correlation between interesting variables.}
}
\value{

}
\description{
batch_pcc() provide a batch way to calculate the partial correlation coefficient between target gene and others when
controlling a third variable
}
\examples{

data("tcga_crc_exp")
data("tcga_crc_purity")
tcga_crc_exp <- t(tcga_crc_exp)
tcga_crc_exp <- data.frame(array = stringr锛氾細str_sub(rownames(tcga_crc_exp), 1, 15), tcga_crc_exp)
dat <- merge(tcga_crc_purity, tcga_crc_exp, by = "array")
TP53cor <- batch_pcc(data = dat, target = "TP53",
feature = colnames(dat)[3:100], method = "pearson", interference = dat$purity)
}
\author{
Rongfang Shen
}
