% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sig_surv_plot.R
\name{batch_sig_surv_plot}
\alias{batch_sig_surv_plot}
\title{Title}
\usage{
batch_sig_surv_plot(
  input_pdata,
  signature,
  ID = "ID",
  column_of_Project = "ProjectID",
  project = NULL,
  time = "time",
  status = "status",
  time_type = "day",
  break_month = "auto",
  palette = "jama",
  cols = NULL,
  mini_sig = "score",
  save_path = "Multiple-KM-plot",
  show_col = TRUE,
  fig.type = "pdf"
)
}
\arguments{
\item{input_pdata}{}

\item{signature}{}

\item{ID}{}

\item{column_of_Project}{}

\item{project}{}

\item{time}{}

\item{status}{}

\item{time_type}{}

\item{break_month}{}

\item{palette}{}

\item{cols}{}

\item{mini_sig}{}

\item{save_path}{}

\item{show_col}{}

\item{fig.type}{}
}
\description{
Title
}
