% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_surv.R
\name{batch_surv}
\alias{batch_surv}
\title{Batch survival analysis on a data frame with multiple variables}
\usage{
batch_surv(pdata, variable, time = "time", status = "status")
}
\arguments{
\item{pdata}{data with follow up data and multiple variables}

\item{variable}{use column c(a:b) to select variables}

\item{time}{survival time}

\item{status}{survival status}
}
\description{
@description  Extract hazard ratio and confidence intervals from a coxph object of subgroup analysis
}
\examples{

sig_surv_result<- batch_surv(pdata = pdata_sig_tme_binary,variable <- c(c(ncol(pdata_stad)+1):ncol(pdata_sig_tme_binary)))
}
\author{
Dongqiang Zeng
}
