% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_surv.R
\name{batch_surv}
\alias{batch_surv}
\title{batch_surv - Batch Survival Analysis Function}
\usage{
batch_surv(
  pdata,
  variable,
  time = "time",
  status = "status",
  best_cutoff = FALSE
)
}
\arguments{
\item{pdata}{The data frame containing the data.}

\item{variable}{The variables to perform survival analysis on.}

\item{time}{The column name representing time. Default is "time".}

\item{status}{The column name representing status. Default is "status".}

\item{best_cutoff}{Whether to use the best cutoff for survival analysis. Default is FALSE. If set to TRUE, the function will calculate hazard ratios and confidence intervals for the binary version of the variables using the best cutoff method.}
}
\value{
A data frame containing hazard ratios and confidence intervals.
}
\description{
This function is used to perform batch survival analysis. It calculates hazard ratios and confidence intervals for the specified variables based on the given data containing time-related information.
}
\examples{

# Loading TCGA-STAD microenvironment signature data
data("sig_stad", package = "IOBR")
# Identifying signatures associated with gastric cancer patient survival
batch_surv(pdata = sig_stad, variable = colnames(sig_stad)[69:ncol(sig_stad)], time = "OS_time", status = "OS_status")
}
\author{
Dongqiang Zeng
}
