% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_wilcoxon.R
\name{batch_wilcoxon}
\alias{batch_wilcoxon}
\title{Batch to conduct wilcoxon test with two groups}
\usage{
batch_wilcoxon(data = data, target = "group", feature = feature)
}
\arguments{
\item{data}{signature matrix with two groups}

\item{target}{name of group: with `high` and `low` category}

\item{feature}{feature used to comparison}
}
\description{
Batch to conduct wilcoxon test with two groups
}
\author{
Dongqiang Zeng
}
