% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_wilcoxon.R
\name{batch_wilcoxon}
\alias{batch_wilcoxon}
\title{Batch to conduct wilcoxon test in two groups}
\usage{
batch_wilcoxon(data, target = "group", group_names = c("High", "Low"), feature)
}
\arguments{
\item{data}{signature matrix with two groups}

\item{target}{name of group:}

\item{group_names}{with `high` and `low` as default}

\item{feature}{feature used to comparison}
}
\value{

}
\description{
Batch to conduct wilcoxon test in two groups
}
\author{
Dongqiang Zeng
}
