% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_wilcoxon.R
\name{batch_wilcoxon}
\alias{batch_wilcoxon}
\title{Batch to conduct wilcoxon test in two groups}
\usage{
batch_wilcoxon(
  data,
  target = "group",
  feature = NULL,
  feature_manipulation = FALSE
)
}
\arguments{
\item{data}{signature matrix with two groups}

\item{target}{name of group;}

\item{feature}{feature used to comparison}

\item{feature_manipulation}{default is FALSE, if true, variables with NA or infinite value will be removed}
}
\description{
Batch to conduct wilcoxon test in two groups
}
\examples{
res<-batch_wilcoxon(data = pdata_sig_tme, target = "Gender", feature = colnames(pdata_sig_tme)[300:400])
}
\author{
Dongqiang Zeng
}
