% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cutoff.R
\name{best_cutoff}
\alias{best_cutoff}
\title{Extract best cut-off and add new binary object to data frame}
\usage{
best_cutoff(pdata, variable, time = "time", status = "status", PrintResult = T)
}
\arguments{
\item{pdata}{phenotype data with survival information and features}

\item{variable}{The name of the continuous variable in the input data for which the best cutoff needs to be determined.}

\item{time}{The name of the column in the input data representing the time-to-event (survival time). The default value is "time".}

\item{status}{The name of the column in the input data representing the event status (censoring information). The default value is "status".}

\item{PrintResult}{A logical value indicating whether to print the results. The default value is TRUE.}
}
\value{
pdata with binary variables
}
\description{
The "best_cutoff" function is used to find the best cutoff point for a continuous variable in survival analysis. It takes input data ("pdata") containing a continuous variable ("variable") and survival information ("time" and "status"). It returns the modified input data with a new binary variable created based on the best cutoff point.
}
\examples{

# Loading TCGA-STAD microenvironment signature data
data("sig_stad", package = "IOBR")
# Finding the best cutoff value of TMEscore for survival
sig_stad2 <- best_cutoff(pdata = sig_stad, variable = "TMEscore_CIR", time = "OS_time",status = "OS_status", PrintResult = T)
table(sig_stad2$TMEscore_CIR_binary)
}
\author{
Dongqiang Zeng
}
