% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cutoff.R
\name{best_cutoff}
\alias{best_cutoff}
\title{Extract best cut-off and add new binary object to data frame}
\usage{
best_cutoff(pdata, variable, time = "time", status = "status", PrintResult = T)
}
\arguments{
\item{pdata}{phenotype data with surcival information and features}

\item{variable}{selected featrues that will be classified into two groups}

\item{time}{column name of survival time}

\item{status}{column name of event of follow up}

\item{PrintResult}{logit, print results of survival analysis before and after classifying.}
}
\value{
pdata with binary variables
}
\description{
Extract best cut-off and add new binary object to data frame
}
\examples{

}
