% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cutoff2.R
\name{best_cutoff2}
\alias{best_cutoff2}
\title{Extract best cut-off and add new binary object to data frame}
\usage{
best_cutoff2(
  pdata,
  variable,
  time = "time",
  status = "status",
  PrintResult = T
)
}
\arguments{
\item{pdata}{phenotype data with surcival information and features}

\item{variable}{selected featrues that will be classified into two groups}

\item{time}{column name of survival time}

\item{status}{column name of event of follow up}

\item{PrintResult}{logit, print results of survival analysis before and after classifying.}
}
\value{
pdata with binary variables
}
\description{
Extract best cut-off and add new binary object to data frame
}
\examples{

}
