% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModel.R
\name{BinomialModel}
\alias{BinomialModel}
\title{Binomial Model construction}
\usage{
BinomialModel(
  x,
  y,
  seed = "123456",
  scale = TRUE,
  train_ratio = 0.7,
  nfold = 10,
  plot = T
)
}
\arguments{
\item{x}{data.frame contains sample ID and features; The first column of x is the sample ID.}

\item{y}{data.frame whose sample ID in the first column and the outcome of each sample in the second sample. Outcome value can be numeric or factor vector.}

\item{seed}{default 123456}

\item{scale}{A logistic: should the x be scaled, default is TRUE.}

\item{train_ratio}{Value between 0-1, eg: 0.7; The ratio is used to split the x and y into training and testing data.}

\item{nfold}{default 10}

\item{plot}{A logistic, default is TRUE.}
}
\value{
a list contain the results of 2 model (Lasso, Ridge) and the input train data.
}
\description{
Binomial Model construction
}
