% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModel.R
\name{BinomialModel}
\alias{BinomialModel}
\title{Traning a binomial model}
\usage{
BinomialModel(
  x,
  y,
  seed = "123456",
  scale = TRUE,
  train_ratio = 0.7,
  nfold = 10
)
}
\arguments{
\item{x}{input matrix or data.frame where samples are in rows and features in columns; The first column of x is the sample ID of which column is "ID".}

\item{y}{input matrix or data.frame whose sample ID in the first column and the outcome of each sample in the second sample. Outcome value can be numeric or factor vector.}

\item{seed}{default 123456}

\item{scale}{A logistic: should the x be scaled, default is TRUE.}

\item{train_ratio}{Value between 0 -1; The ratio was used to split the x and y into training and testing data.}

\item{nfold}{}
}
\value{
a list contain the results of 3 model (Lasso, Ridge and ElasticNet regression model) and the input train data.
}
\description{
Traning a binomial model
}
\examples{
load("./data/crc_response.rda")
load("./data/tme_combine.rda")
binomial_result <- BinomialModel(x = tme_combine, y = na.omit(response[, c("ID", "Stage")]),
seed = "123456", scale = TRUE, train_ratio = 0.7, nfold = 10)
}
