% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score.R
\name{calculate_sig_score}
\alias{calculate_sig_score}
\title{Calculating signature score on a gene expression dataset}
\usage{
calculate_sig_score(
  pdata = NULL,
  eset,
  signature = signature_collection,
  method = "pca",
  mini_gene_count = 3,
  column_of_sample = "ID",
  print_gene_propotion = FALSE,
  adjust_eset = FALSE,
  parallel.size = 1L,
  print_filtered_signatures = FALSE,
  ...
)
}
\arguments{
\item{pdata}{phenotype data of input sample}

\item{eset}{normalized  transcriptomic data: normalized (CPM, TPM, RPKM, FPKM, etc.)}

\item{signature}{List of gene signatures;
such as `signature_tme`, `signature_metabolism`,`signature_collection`, `go_bp`,`kegg`,`hallmark`}

\item{method}{he methods currently supported are `pca`, `ssgsea`, `zscore`,`integration`}

\item{mini_gene_count}{filter out signatures with genes less than minimal gene in expression set;
default is 2 for PCA and z score function}

\item{column_of_sample}{Defines in which column of the `pdata` the identifier of the sample can be found.}

\item{print_gene_propotion}{logical, print the proportion of signature genes in gene matrix}

\item{adjust_eset}{default is FALSE}

\item{parallel.size}{default is 1}

\item{print_filtered_signatures}{logical, print filtered signatures has gene count less than minimal gene count}
}
\value{
data frame with pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating signature score on a gene expression dataset
}
\examples{
# Loading TCGA-STAD expresion data(raw count matrix)
data("eset_stad", package = "IOBR")
# transform count data to tpm
eset <- count2tpm(eset_stad, idType = "ensembl")
# signature score estimation using PCA, z-score, or ssgsea method
calculate_sig_score(eset = eset, signature = signature_tme, method = "pca")

}
\references{
1. Hänzelmann S, Castelo R, Guinney J (2013). “GSVA: gene set variation analysis for microarray and RNA-Seq data.” BMC Bioinformatics, 14, 7. doi: 10.1186/1471-2105-14-7

2. Mariathasan, S., Turley, S., Nickles, D. et al. TGFβ attenuates tumour response to PD-L1 blockade by contributing to exclusion of T cells. Nature 554, 544–548 (2018).
}
\author{
Dongqiang Zeng
}
