% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score.R
\name{calculate_sig_score_integration}
\alias{calculate_sig_score_integration}
\title{Calculating signature score using Integration method}
\usage{
calculate_sig_score_integration(
  pdata = NULL,
  eset,
  signature,
  mini_gene_count = 2,
  column_of_sample = "ID",
  adjust_eset = FALSE,
  parallel.size = 1L
)
}
\arguments{
\item{pdata}{phenotype data of input sample;
if phenotype data is NULL, create a data frame with `Index` and `ID` contain column names of eset}

\item{eset}{normalizaed  transcriptomic data: normalized (CPM, TPM, RPKM, FPKM, etc.)}

\item{signature}{List of gene signatures}

\item{mini_gene_count}{filter out signatures with genes less than minimal gene in expression set; default is 5;
the minimal gene count for ssGSEA methods should larger than 5 for the robustness of the calculation}

\item{column_of_sample}{Defines in which column of pdata the sample identifier can be found.}

\item{adjust_eset}{default is FALSE, if true, data with Inf or zero sd will be replaced}

\item{parallel.size}{default is 1}
}
\value{
data frame with pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating signature score using Integration method
}
\examples{
# Loading TCGA-STAD expresion data(raw count matrix)
data("eset_stad", package = "IOBR")
# transform count data to tpm
eset <- count2tpm(eset_stad, idType = "ensembl")
# signature score estimation using PCA, z-score, and ssgsea method
calculate_sig_score_integration(eset = eset, signature = signature_tme)
}
\author{
Dongqiang Zeng
}
