% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score.R
\name{calculate_sig_score_pca}
\alias{calculate_sig_score_pca}
\title{Calculating signature score using PCA method}
\usage{
calculate_sig_score_pca(
  pdata,
  eset,
  signature,
  mini_gene_count,
  column_of_sample
)
}
\arguments{
\item{pdata}{phenotype data of input sample;
if phenotype data is NULL, create a data frame with `Index` and `ID` contain column names of eset}

\item{eset}{normalizaed  transcriptomic data: normalized (CPM, TPM, RPKM, FPKM, etc.)}

\item{signature}{List of gene signatures;}

\item{mini_gene_count}{filter out signatures with genes less than minimal gene in expression set}

\item{column_of_sample}{Defines in which column of pdata the sample identifier can be found.}
}
\value{
matrix;pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating signature score using PCA method
}
\author{
Dongqiang Zeng
}
