% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{CalculateTimeROC}
\alias{CalculateTimeROC}
\title{Calculate Time ROC}
\usage{
CalculateTimeROC(model, newx, s, acture.y, foldername, modelname)
}
\arguments{
\item{modelname}{}
}
\value{

}
\description{
Calculate Time ROC
}
