% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_bar_plot.R
\name{cell_bar_plot}
\alias{cell_bar_plot}
\title{Visualize the results as a stacked bar chart with tidyverse/ggplot2.}
\usage{
cell_bar_plot(
  input,
  id = "ID",
  title = "Cell Fraction",
  legend.position = "bottom",
  coord_filp = TRUE,
  palette = 3,
  show_col = F,
  cols = NULL
)
}
\arguments{
\item{input}{input  result of `CIBERSORT` or `quantiseq`}

\item{id}{patient identifier of input}

\item{title}{plot title}

\item{legend.position}{legend position}

\item{coord_filp}{logical variables, `coord_filp` of ggplot2}

\item{palette}{default is palette3}

\item{show_col}{default is FALSE}

\item{cols}{default is NULL, user can define colors manually}
}
\description{
Visualize the results as a stacked bar chart with tidyverse/ggplot2.
}
\examples{
res<-cell_bar_plot(input = cibersort[1:20,],id = "ID")

}
