% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_bar_plot.R
\name{cell_bar_plot}
\alias{cell_bar_plot}
\title{Visualize the results as a stacked bar chart with tidyverse/ggplot2.}
\usage{
cell_bar_plot(
  input,
  id = "ID",
  title = "Cell Fraction",
  features = NULL,
  pattern = NULL,
  legend.position = "bottom",
  coord_filp = TRUE,
  palette = 3,
  show_col = F,
  cols = NULL
)
}
\arguments{
\item{input}{input  result of `CIBERSORT` or `quantiseq`}

\item{id}{patient identifier of input}

\item{title}{plot title}

\item{features}{default is NULL, user can define which column to draw}

\item{pattern}{patterns of features}

\item{legend.position}{legend position}

\item{coord_filp}{logical variables, `coord_filp` of ggplot2}

\item{palette}{default is palette3}

\item{show_col}{default is FALSE}

\item{cols}{default is NULL, user can define colors manually}
}
\value{
A ggplot2 object representing the bar plot of cell fractions.
}
\description{
cell_bar_plot() provides batch visualization of tumor microenvironment (TME) cell fractions. It supports input from deconvolution results generated by methodologies such as 'CIBERSORT', 'EPIC', and 'quanTIseq'. The function enables comparison of TME cell distributions within one sample or among different samples.
}
\examples{
# Loading TCGA-STAD microenvironment data
data("sig_stad", package = "IOBR")
# showing 20 tumor microenvironment cell proportion deconvoluted by CIBERSORT algorithm
cell_bar_plot(input = sig_stad[1:20, ], id = "ID", features = colnames(sig_stad)[25:46])
}
