% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eset.R
\name{check_eset}
\alias{check_eset}
\title{Check the value of gene expression set}
\usage{
check_eset(eset, print_result = FALSE, estimate_sd = FALSE)
}
\arguments{
\item{eset}{gene expression set}

\item{print_result}{logical, default is FALSE}

\item{estimate_sd}{sd estimation}
}
\description{
Check the value of gene expression set
}
