% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{clear_iobr_cache}
\alias{clear_iobr_cache}
\title{Clear IOBR Data Cache}
\usage{
clear_iobr_cache()
}
\value{
Invisible NULL. Called for side effects of clearing the cache.
}
\description{
Removes all cached data files downloaded from GitHub.
}
