% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pd_eset.R
\name{combine_pd_eset}
\alias{combine_pd_eset}
\title{Title}
\usage{
combine_pd_eset(
  eset,
  pdata,
  id_pdata = "ID",
  feas = NULL,
  feature_manipulation = T,
  scale = T,
  choose_who_when_duplicate = "eset"
)
}
\arguments{
\item{eset}{expression set}

\item{pdata}{phenotype data}

\item{id_pdata}{identifier of pdata which could be matched with column name of eset}

\item{feas}{selected features of eset to combine with pdata}

\item{feature_manipulation}{default is T, data with outliers will be filtered}

\item{scale}{default is true, scaling data}

\item{choose_who_when_duplicate}{when duplicated variables exist, preserve data of pdata or eset.}
}
\description{
Title
}
