% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pd_eset.R
\name{combine_pd_eset}
\alias{combine_pd_eset}
\title{combine pdata and eset}
\usage{
combine_pd_eset(
  eset,
  pdata,
  id_pdata = "ID",
  feas = NULL,
  feature_manipulation = TRUE,
  scale = TRUE,
  choose_who_when_duplicate = "eset"
)
}
\arguments{
\item{eset}{Expression set matrix or data frame.}

\item{pdata}{Phenotype data.}

\item{id_pdata}{String indicating the column name in pdata that corresponds to the eset column.}

\item{feas}{Vector of feature names to include in the combined dataset. Default is NULL, which includes all features.}

\item{feature_manipulation}{Logical value indicating whether to perform feature manipulation on the data. Default is TRUE.}

\item{scale}{Logical value indicating whether to scale the expression set. Default is TRUE.}

\item{choose_who_when_duplicate}{String indicating the preference when encountering duplicate IDs in pdata and eset.}
}
\value{
A data frame combining phenotype data and expression set.
}
\description{
combine_pd_eset combines the expression set (eset) with phenotype data (pdata).
}
\examples{
# Loading TCGA-STAD expresion data(raw count matrix)
data("eset_stad", package = "IOBR")
# transform count data to tpm
eset <- count2tpm(eset_stad, idType = "ensembl")
colnames(eset) <- substring(colnames(eset), 1, 12)
# Loading TCGA-STAD microenvironment signature data
data("sig_stad", package = "IOBR")
eset <- count2tpm(eset_stad)
colnames(eset) <- substring(colnames(eset), 1, 12)
input <- combine_pd_eset(eset = eset, pdata = sig_stad, feas = unique(unlist(signature_collection)))
}
