% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{ConvertRownameToLoci}
\alias{ConvertRownameToLoci}
\title{Convert Rowname To Loci}
\usage{
ConvertRownameToLoci(cancerGeneExpression)
}
\arguments{
\item{cancerGeneExpression}{}
}
\value{

}
\description{
Convert Rowname To Loci
}
