% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{CoreAlg}
\alias{CoreAlg}
\title{Perform Nu-Regression Using Support Vector Machines}
\usage{
CoreAlg(X, y, absolute, abs_method)
}
\arguments{
\item{X}{A matrix or data frame containing the predictor variables.}

\item{y}{A numeric vector containing the response variable.}

\item{absolute}{Logical indicating whether to use absolute space for weights. Default is FALSE.}

\item{abs_method}{String specifying the method for absolute space weights: "sig.score" or "no.sumto1".}
}
\value{
A list containing the weights (`w`), root mean squared error (`mix_rmse`), and correlation coefficient (`mix_r`).
}
\description{
This function performs nu-regression using support vector machines (SVM) and calculates weights,
root mean squared error (RMSE), and correlation coefficient (R).
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
y <- rnorm(10)
result <- CoreAlg(X, y, absolute = FALSE, abs_method = "sig.score")

}
