% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2tpm.R
\docType{methods}
\name{count2tpm}
\alias{count2tpm}
\title{Convert read counts to transcripts per million (TPM)}
\usage{
count2tpm(countMat, idType = "Ensembl", org = "hsa")
}
\arguments{
\item{countMat}{A read count matrix, with geneid as rownames and sample as columns.}

\item{idType}{Type of gene id. "Ensembl" "ENTREZ","SYMBOL",}

\item{org}{Organism, hsa or mmu.}
}
\value{
A tpm expression profile.
}
\description{
Convert read counts to transcripts per million (TPM)
}
\author{
Wubing Zhang

Dongqiang Zeng
}
