% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2tpm2.R
\docType{methods}
\name{count2tpm}
\alias{count2tpm}
\title{Convert read counts to transcripts per million (TPM)}
\usage{
count2tpm(
  countMat,
  idType = "Ensembl",
  org = "hsa",
  source = "local",
  effLength = NULL,
  id = "id",
  gene_symbol = "symbol",
  length = "eff_length",
  remove_redundancy = "mean"
)
}
\arguments{
\item{countMat}{A read count matrix, with `geneid` as row names and samples as columns.}

\item{idType}{Type of gene id. "Ensembl" "ENTREZ","SYMBOL"}

\item{org}{Organism, `hsa` or `mmus`}

\item{source}{default is "local". Other option is `biomart`. user can also provide `effLength` manually, if `idType` is `ensembl`, and source is set to `local`, `effLength` was provided by IOBR which was estimated by function `getGeneLengthAndGCContent` of EDASeq package at 2023-02-10.}

\item{effLength}{user can provide gene Length data with gene id and effective length}

\item{id}{id matched to row names of expression set}

\item{gene_symbol}{column name of gene symbol}

\item{length}{column name of gene length}

\item{remove_redundancy}{default is mean, other options are `sd` or `median`}
}
\value{
A tpm expression profile.
}
\description{
Convert read counts to transcripts per million (TPM)
}
\author{
Wubing Zhang

Dongqiang Zeng
}
