% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_cibersort}
\alias{deconvo_cibersort}
\title{DeconvolVe immune microenvironment using CIBERSORT}
\usage{
deconvo_cibersort(eset, project, arrays, absolute = FALSE, perm = 1000)
}
\arguments{
\item{eset}{expression set with gene symbol at rowname, sample ID at column}

\item{project}{project name used to distinguish different datasets}

\item{arrays}{logical: Runs methods in a mode optimized for microarray data.}

\item{absolute}{logical: Runs CIBERSORT in absolute mode}

\item{perm}{permutation to run CIBERSORT}
}
\value{
cibersrot with immune cell fractions
}
\description{
CIBERSORT is only freely available to academic users.
A license an the binary can be obtained from https://cibersort.stanford.edu.
}
\examples{
cibersort_result<-deconvo_cibersort(eset = eset,project = "ACRG",arrays = TRUE,absolute = FALSE, perm = 500)
}
\author{
Dongqiang Zeng
}
