% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_cibersort}
\alias{deconvo_cibersort}
\title{Decoding immune microenvironment using CIBERSORT}
\usage{
deconvo_cibersort(
  eset,
  project = NULL,
  arrays,
  perm = 1000,
  absolute = FALSE,
  abs_method = "sig.score"
)
}
\arguments{
\item{eset}{expression set with gene symbol at row name, sample ID at column}

\item{project}{project name used to distinguish different data sets}

\item{arrays}{logical: Runs methods in a mode optimized for microarray data.}

\item{perm}{permutation to run CIBERSORT}

\item{absolute}{logical: Runs CIBERSORT in absolute mode}
}
\value{
cibersrot with immune cell fractions
}
\description{
CIBERSORT is only freely available to academic users.
A license an the binary can be obtained from https://cibersort.stanford.edu.
}
\examples{
# Loading TCGA-STAD expresion data(raw count matrix)
data(eset_stad, package = "IOBR")
eset <- count2tpm(countMat = eset_stad, source = "local", idType = "ensembl")
cibersort_result <- deconvo_cibersort(eset = eset, project = "TCGA-STAD", arrays = FALSE, absolute = FALSE, perm = 500)
}
\author{
Dongqiang Zeng
}
