% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_epic}
\alias{deconvo_epic}
\title{DeconvolVe immune microenvironment using EPIC: FOR RANseq mostly}
\usage{
deconvo_epic(eset, project, tumor)
}
\arguments{
\item{eset}{expression set with genes at row, sample ID at column}

\item{project}{project name used to distinguish different datasets}

\item{tumor}{is input sample tumor or normal}
}
\value{
EPIC with immune cell fractions
}
\description{
DeconvolVe immune microenvironment using EPIC: FOR RANseq mostly
}
\examples{
epic_result<-deconvo_epic(eset = eset,project = "GSE62254",tumor = TRUE)

}
\author{
Dongqiang Zeng
}
