% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_epic}
\alias{deconvo_epic}
\title{Deconvolve Immune Microenvironment Using EPIC}
\usage{
deconvo_epic(eset, project = NULL, tumor)
}
\arguments{
\item{eset}{A gene expression matrix with genes in rows and samples in columns.}

\item{project}{Optional project name to distinguish different datasets.}

\item{tumor}{Logical indicating whether the input samples are tumor (TRUE) or normal (FALSE).}
}
\value{
A data frame containing EPIC immune cell fraction estimates with sample IDs.
}
\description{
This function estimates immune cell fractions in the tumor microenvironment using the EPIC algorithm,
primarily designed for RNA-seq data.
}
\examples{
# Load TCGA-STAD expression data (raw count matrix)
data("eset_stad", package = "IOBR")
eset <- count2tpm(countMat = eset_stad, source = "local", idType = "ensembl")
epic_result <- deconvo_epic(eset = eset, project = "TCGA-STAD", tumor = TRUE)

}
\author{
Dongqiang Zeng
}
