% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_estimate}
\alias{deconvo_estimate}
\title{Calculation of stromal, immune, and ESTIMATE scores}
\usage{
deconvo_estimate(eset, project = NULL, platform = "affymetrix")
}
\arguments{
\item{eset}{expression set with genes at row, sample ID at column}

\item{project}{project name used to distinguish different data sets}

\item{platform}{character string indicating platform type. Defaults to "affymetrix"}
}
\description{
Calculation of stromal, immune, and ESTIMATE scores
}
\author{
Dongqiang Zeng
}
