% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_estimate}
\alias{deconvo_estimate}
\title{Calculation of stromal, immune, and ESTIMATE scores}
\usage{
deconvo_estimate(eset, project = NULL, platform = "affymetrix")
}
\arguments{
\item{eset}{expression set with genes at row, sample ID at column}

\item{project}{project name used to distinguish different data sets}

\item{platform}{character string indicating platform type. Defaults to "affymetrix"}
}
\value{
A data frame with the calculated stromal, immune, and ESTIMATE scores.
}
\description{
Calculation of stromal, immune, and ESTIMATE scores
}
\examples{
# Loading TCGA-STAD expresion data(raw count matrix)
data(eset_stad, package = "IOBR")
eset <- count2tpm(countMat = eset_stad, source = "local", idType = "ensembl")
deconvo_estimate(eset)
}
\author{
Dongqiang Zeng
}
