% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_ips}
\alias{deconvo_ips}
\title{Calculate Immunophenoscore (IPS)}
\usage{
deconvo_ips(eset, project = NULL, plot = FALSE)
}
\arguments{
\item{eset}{Gene expression matrix.}

\item{project}{Optional project name. Default is `NULL`.}

\item{plot}{Logical: generate visualization. Default is `FALSE`.}
}
\value{
Data frame with IPS scores. Columns suffixed with `_IPS`.
}
\description{
Calculates immune phenotype scores from gene expression data.
}
\examples{
\donttest{
eset_stad <- load_data("eset_stad")
anno_grch38 <- load_data("anno_grch38")
eset <- anno_eset(eset = eset_stad, annotation = anno_grch38, probe = "id")
ips_result <- deconvo_ips(eset = eset, project = "TCGA-STAD")
}
}
\author{
Dongqiang Zeng
}
