% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_ips}
\alias{deconvo_ips}
\title{DeconvolVe immune microenvironment using IPS}
\usage{
deconvo_ips(eset, project, plot)
}
\arguments{
\item{eset}{expression set with genes at row, sample ID at column}

\item{project}{project name used to distinguish different datasets}
}
\value{
IPS data frame
}
\description{
DeconvolVe immune microenvironment using IPS
}
\examples{
ips_result<-deconvo_ips(eset = eset_ec,project = "GSE62254")

}
\author{
Dongqiang Zeng
}
