% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_mcpcounter}
\alias{deconvo_mcpcounter}
\title{Deconvolve Immune Microenvironment Using MCP-Counter}
\usage{
deconvo_mcpcounter(eset, project = NULL)
}
\arguments{
\item{eset}{Gene expression matrix with HGNC symbols as row names.}

\item{project}{Optional project name. Default is `NULL`.}
}
\value{
Data frame with MCP-counter scores. Columns suffixed with
  `_MCPcounter`.
}
\description{
Estimates immune cell abundances using MCP-counter.
}
\examples{
\donttest{
eset_stad <- load_data("eset_stad")
eset <- count2tpm(eset_stad, idType = "ensembl")
mcp_result <- deconvo_mcpcounter(eset = eset, project = "TCGA-STAD")
}
}
\author{
Dongqiang Zeng
}
