% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_mcpcounter}
\alias{deconvo_mcpcounter}
\title{Estimating immune microenvironment using MCP-counter}
\usage{
deconvo_mcpcounter(eset, project = NULL)
}
\arguments{
\item{eset}{expression set with genes at row, sample ID at column}

\item{project}{project name used to distinguish different datasets}
}
\value{
MCP-counter with immune cell fractions
}
\description{
Estimating immune microenvironment using MCP-counter
}
\examples{
mcp_result<-deconvo_mcpcounter(eset = eset_stad,project = "TCGA-STAD")

}
\author{
Dongqiang Zeng
}
