% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_quantiseq}
\alias{deconvo_quantiseq}
\title{Deconvolve Using quanTIseq}
\usage{
deconvo_quantiseq(eset, project = NULL, tumor, arrays, scale_mrna)
}
\arguments{
\item{eset}{Gene expression matrix.}

\item{project}{Optional project name. Default is `NULL`.}

\item{tumor}{Logical: tumor samples. Must be specified.}

\item{arrays}{Logical: microarray data. Must be specified.}

\item{scale_mrna}{Logical: correct for mRNA content. Must be specified.}
}
\value{
Data frame with quanTIseq cell fractions. Columns suffixed with
  `_quantiseq`.
}
\description{
quanTIseq deconvolution for RNA-seq immune cell fractions.
}
\examples{
\dontrun{
eset_stad <- load_data("eset_stad")
eset <- count2tpm(eset_stad, idType = "ensembl")
res <- deconvo_quantiseq(
  eset = eset, project = "stad", tumor = TRUE,
  arrays = FALSE, scale_mrna = FALSE
)
}
}
\author{
Dongqiang Zeng
}
