% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_quantiseq}
\alias{deconvo_quantiseq}
\title{Deconvolute using the quanTIseq technique}
\usage{
deconvo_quantiseq(eset, project = NULL, tumor, arrays, scale_mrna)
}
\arguments{
\item{eset}{a m x n matrix with m genes and n samples}

\item{project}{}

\item{tumor}{logistic}

\item{arrays}{microarray}
}
\value{

}
\description{
Deconvolute using the quanTIseq technique
}
