% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_quantiseq}
\alias{deconvo_quantiseq}
\title{Deconvoluting micrornvironment using the quanTIseq technique}
\usage{
deconvo_quantiseq(eset, project = NULL, tumor, arrays, scale_mrna)
}
\arguments{
\item{eset}{gene expression data}

\item{project}{default is NULL, project name used to distinguish different data sets}

\item{tumor}{logistic variable}

\item{arrays}{logistic variable, is data generated from microarray}

\item{scale_mrna}{logistic variable}
}
\value{
A data frame with the estimated cell type fractions, with columns named according to the cell types and suffixed with "_quantiseq".
}
\description{
This function performs deconvolution of the tumor microenvironment using the quanTIseq technique. It estimates the fraction of different cell types in the expression dataset.
}
\examples{
#' # Loading TCGA-STAD expresion data(raw count matrix)
data(eset_stad, package = "IOBR")
eset <- count2tpm(countMat = eset_stad, source = "local", idType = "ensembl")
deconvo_quantiseq(eset = eset, project = "stad", tumor = TRUE, arrays = FALSE, scale_mrna = FALSE)
}
