% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_ref}
\alias{deconvo_ref}
\title{Estimate the fraction of cell types using defined reference genes}
\usage{
deconvo_ref(
  eset,
  project = NULL,
  arrays,
  method = "svm",
  perm,
  reference,
  scale_reference
)
}
\arguments{
\item{eset}{expression data with matched gene id of reference}

\item{arrays}{a logical value. If TRUE, the columns of the input data will be normalized to have the same quantiles.}

\item{method}{deconvolution method. must be "svm" or "lsei"}

\item{perm}{permutation to run CIBERSORT}

\item{reference}{immune cell gene matrix; eg lm22, lm6 or can be generate using generateRef/generateRef_rnaseq}

\item{scale_reference}{a logical value indicating whether the reference be scaled or not. If TRUE, the value in reference file will be centered and scaled in row direction.}
}
\value{

}
\description{
Estimate the fraction of cell types using defined reference genes
}
\author{
Dongqiang Zeng

Rongfang Shen
}
