% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_ref}
\alias{deconvo_ref}
\title{Estimate the fraction of cell types using defined reference genes}
\usage{
deconvo_ref(
  eset,
  project = NULL,
  arrays = TRUE,
  method = "svr",
  perm = 100,
  reference,
  scale_reference = TRUE,
  absolute.mode = FALSE,
  abs.method = "sig.score"
)
}
\arguments{
\item{eset}{expression data with matched gene id of reference}

\item{project}{project name used to distinguish different datasets}

\item{arrays}{a logical value. If TRUE, the columns of the input data will be normalized to have the same quantiles.}

\item{method}{deconvolution method. must be "svr" or "lsei"}

\item{perm}{permutation to run svr}

\item{reference}{immune cell gene matrix; eg lm22, lm6 or can be generate using generateRef/generateRef_rnaseq}

\item{scale_reference}{a logical value indicating whether the reference be scaled or not. If TRUE, the value in reference file will be centered and scaled in row direction.}

\item{absolute.mode}{absolute.mode, default is FALSE}

\item{abs.method}{abs.method, default is sig.score}
}
\value{
A data frame with the estimated cell type fractions. The columns are named with the cell types and suffixed with "_CIBERSORT".
}
\description{
Estimate the fraction of cell types using defined reference genes
}
\examples{
# Loading TCGA-STAD expresion data(raw count matrix)
data(eset_stad, package = "IOBR")
eset <- count2tpm(countMat = eset_stad, source = "local", idType = "ensembl")
deconvo_ref(eset = eset, reference = lm22)
}
\author{
Dongqiang Zeng

Rongfang Shen
}
