% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_timer}
\alias{deconvo_timer}
\title{Deconvoluting using the TIMER technique
Unlike the other methods, TIMER needs the specification of the cancer type for each sample.}
\usage{
deconvo_timer(eset, project = NULL, indications = NULL)
}
\arguments{
\item{eset}{gene matrix}

\item{project}{default is NULL, project name used to distinguish different data sets}

\item{indications}{a n-vector giving and indication string (e.g. 'brca') for each sample. Accepted indications are 'kich', 'blca', 'brca', 'cesc', 'gbm', 'hnsc', 'kirp', 'lgg','lihc', 'luad', 'lusc', 'prad', 'sarc', 'pcpg', 'paad', 'tgct','ucec', 'ov', 'skcm', 'dlbc', 'kirc', 'acc', 'meso', 'thca','uvm', 'ucs', 'thym', 'esca', 'stad', 'read', 'coad', 'chol'}
}
\value{

}
\description{
Deconvoluting using the TIMER technique
Unlike the other methods, TIMER needs the specification of the cancer type for each sample.
}
\examples{

}
