% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_timer}
\alias{deconvo_timer}
\title{Deconvolute using the TIMER technique
Unlike the other methods, TIMER needs the specification of the cancer type for each sample.}
\usage{
deconvo_timer(eset, project = NULL, indications = NULL)
}
\arguments{
\item{eset}{a m x n matrix with m genes and n samples}

\item{project}{default is NULL}

\item{indications}{a n-vector giving and indication string (e.g. 'brca') for each sample. Accepted indications are 'kich', 'blca', 'brca', 'cesc', 'gbm', 'hnsc', 'kirp', 'lgg','lihc', 'luad', 'lusc', 'prad', 'sarc', 'pcpg', 'paad', 'tgct','ucec', 'ov', 'skcm', 'dlbc', 'kirc', 'acc', 'meso', 'thca','uvm', 'ucs', 'thym', 'esca', 'stad', 'read', 'coad', 'chol'}
}
\value{

}
\description{
Deconvolute using the TIMER technique
Unlike the other methods, TIMER needs the specification of the cancer type for each sample.
}
\examples{

}
