% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_tme}
\alias{deconvo_tme}
\title{Deconvoluting Tumor microenvironment on a transcriptomic dataset}
\usage{
deconvo_tme(
  eset,
  project = NULL,
  method = tme_deconvolution_methods,
  arrays = FALSE,
  tumor = TRUE,
  perm = 1000,
  reference,
  scale_reference,
  plot = FALSE,
  scale_mrna,
  group_list = NULL,
  platform = "affymetrix",
  absolute.mode = FALSE,
  abs.method = "sig.score",
  ...
)
}
\arguments{
\item{eset}{A gene expression matrix
Either: A numeric matrix or data.frame with HGNC gene symbols as rownames and sample identifiers as colnames. In both cases, data must be on non-log scale.}

\item{project}{project name used to distinguish different data sets, default is NULL}

\item{method}{a string specifying the method.
Supported methods are `mcp_counter`, `epic`, `xcell`, `cibersort`, `cibersort_abs`, `ips`, `quantiseq`, `estimate`,`timer`, `svm`,`lsei`，`timer`, `quantiseq`.}

\item{arrays}{Runs methods in a mode optimized for microarray data.
Currently affects `CIBERSORT`, `svm` and `xCell`.}

\item{tumor}{logical. use a signature matrix/procedure optimized for tumor samples,
if supported by the method. Currently affects `EPIC`}

\item{perm}{set permutations for statistical analysis (≥100 permutations recommended).
Currently affects `CIBERSORT` and `svm_ref`}

\item{reference}{immune cell gene matrix; eg lm22, lm6 or can be generate using generateRef/generateRef_rnaseq}

\item{scale_reference}{a logical value indicating whether the reference be scaled or not. If TRUE, the value in reference file will be centered and scaled in row direction. Currently affects `svm` and `lsei` method}

\item{plot}{Currently affects `IPS` method}

\item{scale_mrna}{logical. If FALSE, disable correction for mRNA content of different cell types.
This is supported by methods that compute an absolute score (EPIC and quanTIseq)}

\item{group_list}{tumor type list of samples}

\item{platform}{character string indicating platform type. Defaults to "affymetrix"
Currently affects `ESTIMATE` method}

\item{absolute.mode}{Run CIBERSORT or SVM in absolute mode (default = FALSE)}

\item{abs.method}{if absolute is set to TRUE, choose method: 'no.sumto1' or 'sig.score'}

\item{...}{arguments passed to the respective method}
}
\value{
`data.frame` with `ID` as first column and other column with the
    calculated cell fractions for each sample.
}
\description{
Deconvoluting Tumor microenvironment on a transcriptomic dataset
}
\examples{

}
\author{
Dongqiang Zeng

Rongfang Shen
}
