% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_xcell}
\alias{deconvo_xcell}
\title{Decoding immune microenvironment using xCell}
\usage{
deconvo_xcell(eset, project = NULL, arrays = FALSE)
}
\arguments{
\item{eset}{expression set with genes at row, sample ID at column}

\item{project}{project name used to distinguish different data sets}

\item{array}{transcrptomic data type}
}
\value{
xCell with immune cell fractions
}
\description{
Decoding immune microenvironment using xCell
}
\examples{
# Loading TCGA-STAD expresion data(raw count matrix)
data(eset_stad, package = "IOBR")
eset <- count2tpm(countMat = eset_stad, source = "local", idType = "ensembl")
xcell_result<-deconvo_xcell(eset = eset, project = "TCGA-STAD", arrays = FALSE)

}
\author{
Dongqiang Zeng
}
