% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\name{deconvo_xcell}
\alias{deconvo_xcell}
\title{DeconvolVe immune microenvironment using xCell}
\usage{
deconvo_xcell(eset, project = NULL, arrays)
}
\arguments{
\item{eset}{expression set with genes at row, sample ID at column}

\item{project}{project name used to distinguish different datasets}

\item{array}{transcrptomic data type}
}
\value{
xCell with immune cell fractions
}
\description{
DeconvolVe immune microenvironment using xCell
}
\examples{
xcell_result<-deconvo_xcell(eset = eset_ec,project = "GSE62254")

}
\author{
Dongqiang Zeng
}
