% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{deconvolute_timer.default}
\alias{deconvolute_timer.default}
\title{deconvolute tumor microenvironment using timer}
\usage{
deconvolute_timer.default(args)
}
\arguments{
\item{args}{environment}
}
\value{

}
\description{
deconvolute tumor microenvironment using timer
}
