% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_mytheme.R
\name{design_mytheme}
\alias{design_mytheme}
\title{Design my ggplpt2 theme}
\usage{
design_mytheme(
  plot_title_size = 2,
  axis_title_size = 2,
  axis_text_size = 12,
  axis_angle = 60,
  hjust = 1,
  legend.position = "bottom",
  legend.direction = "horizontal"
)
}
\arguments{
\item{plot_title_size}{title size}

\item{axis_title_size}{axis title size}

\item{axis_text_size}{axis text size}

\item{axis_angle}{axis angle}

\item{hjust}{horizontal justification}
}
\value{
mytheme
}
\description{
Design my ggplpt2 theme
}
\author{
Dongqiang Zeng
}
