% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_mytheme.R
\name{design_mytheme}
\alias{design_mytheme}
\title{design_mytheme}
\usage{
design_mytheme(
  theme = "light",
  plot_title_size = 2,
  axis_title_size = 2,
  axis_text_size = 12,
  axis_angle = 60,
  hjust = 1,
  legend.position = NULL,
  legend.direction = NULL,
  legend.size = 0.25,
  legend.key.height = 0.5,
  legend.key.width = 0.5,
  legend.size.text = 10,
  legend.box = "horizontal"
)
}
\arguments{
\item{theme}{The desired theme for the plot. Options include "light", "bw", "classic", and "classic2". Default is "light".}

\item{plot_title_size}{The size of the plot title. Default is 2.}

\item{axis_title_size}{The size of the axis titles. Default is 2.}

\item{axis_text_size}{The size of the axis tick mark labels. Default is 12.}

\item{axis_angle}{The angle of rotation for the x-axis tick mark labels. Default is 60.}

\item{hjust}{horizontal justification}

\item{legend.position}{The position of the legend in the plot. Options include "none", "left", "right", "bottom", and "top". Default is NULL.}

\item{legend.direction}{The direction of the legend items. Options include "horizontal" and "vertical". Default is NULL.}

\item{legend.size}{The size of the legend key. Default is 0.25.}

\item{legend.key.height}{The height of the legend key. Default is 0.5.}

\item{legend.key.width}{The width of the legend key. Default is 0.5.}

\item{legend.size.text}{The size of the legend text labels. Default is 10.}

\item{legend.box}{The orientation of the legend box. Options include "horizontal" and "vertical". Default is "horizontal".}
}
\value{
mytheme
}
\description{
This function modifies a ggplot2 theme based on user-defined parameters for various plot elements. It allows customization of the theme's appearance, including title size, axis size, legend position, and more.
}
\examples{
mytheme <- design_mytheme()
}
\author{
Dongqiang Zeng
}
