% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_mytheme.R
\name{design_mytheme}
\alias{design_mytheme}
\title{Design my ggplpt2 theme}
\usage{
design_mytheme(
  theme = "light",
  plot_title_size = 2,
  axis_title_size = 2,
  axis_text_size = 12,
  axis_angle = 60,
  hjust = 1,
  legend.position = "bottom",
  legend.direction = "horizontal",
  legend.size = 12
)
}
\arguments{
\item{theme}{default is light}

\item{plot_title_size}{title size}

\item{axis_title_size}{axis title size}

\item{axis_text_size}{axis text size}

\item{axis_angle}{axis angle}

\item{hjust}{horizontal justification}

\item{legend.position}{legend.position default is bottom}

\item{legend.direction}{legend.direction horizontal or vertical}

\item{legend.size}{font size of legend}
}
\value{
mytheme
}
\description{
Design my ggplpt2 theme
}
\examples{
mytheme <- design_mytheme()
}
\author{
Dongqiang Zeng
}
