% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{doPerm}
\alias{doPerm}
\title{Title}
\usage{
doPerm(perm, X, Y, absolute, abs_method)
}
\arguments{
\item{perm}{Number of permutations to perform.}

\item{X}{Input matrix or data frame containing the predictor variables.}

\item{Y}{Numeric vector containing the response variable.}

\item{absolute}{Logical value indicating whether to use absolute space or relative space for the weights.}

\item{abs_method}{String indicating the method to calculate the weights in absolute space. Can be either 'sig.score' or 'no.sumto1'.}
}
\value{
A list containing the distribution of correlation coefficients from the permutations.
}
\description{
doPerm performs permutation-based sampling and runs the CoreAlg function iteratively.
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
y <- rnorm(10)
result <- doPerm(1000, X, Y, absolute = FALSE, abs_method = "sig.score")

}
