% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_distribution.R
\name{eset_distribution}
\alias{eset_distribution}
\title{Title}
\usage{
eset_distribution(eset, quantile = 3, log = TRUE, project = NULL)
}
\arguments{
\item{project}{}
}
\description{
Title
}
\examples{
画图--箱图
画图--密度曲线图
}
