% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_helper.R
\name{estimateScore}
\alias{estimateScore}
\title{estimateScore}
\usage{
estimateScore(
  input.ds,
  output.ds,
  platform = c("affymetrix", "agilent", "illumina")
)
}
\arguments{
\item{platform}{"affymetrix", "agilent", "illumina"}
}
\description{
estimateScore
}
