% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_cor.R
\name{exact_pvalue}
\alias{exact_pvalue}
\title{Calculate exact p value of correlation}
\usage{
exact_pvalue(x, y, method)
}
\arguments{
\item{x}{variables}

\item{y}{variables}

\item{method}{method used to conduct correlation analysis}
}
\value{
A numeric value representing the exact p-value of the correlation.
}
\description{
This function calculates the exact p-value for the correlation between two variables using a specified method.
}
\examples{

# Loading TCGA-STAD microenvironment signature data
data("sig_stad", package = "IOBR")
exact_pvalue(sig_stad$CD8.T.cells, sig_stad$CD_8_T_effector, method = "spearman")
}
