% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_cor.R
\name{exact_pvalue}
\alias{exact_pvalue}
\title{Calculate exact p value of correlation}
\usage{
exact_pvalue(x, y, method)
}
\arguments{
\item{x}{variables}

\item{y}{variables}

\item{method}{method used to conduct correlation analysis}
}
\description{
Calculate exact p value of correlation
}
