% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_manipulation.R
\name{feature_manipulation}
\alias{feature_manipulation}
\title{Filtered features with multiple methods}
\usage{
feature_manipulation(data = data, feature = feature)
}
\arguments{
\item{data}{data with features in the columns}

\item{feature}{features with NA or Infinity value}
}
\value{
filtered features
}
\description{
Filtered features with multiple methods
}
\author{
Dongqiang Zeng
}
