% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_manipulation.R
\name{feature_manipulation}
\alias{feature_manipulation}
\title{Filtering features using multiple methods}
\usage{
feature_manipulation(
  data,
  feature = NULL,
  is_matrix = FALSE,
  print_result = FALSE
)
}
\arguments{
\item{data}{data with features in the columns}

\item{feature}{features with NA or Infinity value}

\item{is_matrix}{if input is a matrix, data will be transformed and feature will be row names of original data}

\item{print_result}{logical variables, if TRUE, feature detail will be printed in the console}
}
\value{
filtered features
}
\description{
Filtering features using multiple methods
}
\author{
Dongqiang Zeng
}
