% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_manipulation.R
\name{feature_manipulation}
\alias{feature_manipulation}
\title{Manipulate Features in a Matrix or Data Frame}
\usage{
feature_manipulation(data, feature, is_matrix = FALSE, print_result = FALSE)
}
\arguments{
\item{data}{A matrix or data frame containing the features to be manipulated.}

\item{feature}{A vector of feature names to manipulate. Ignored if `is_matrix` is TRUE.}

\item{is_matrix}{Logical indicating whether the input data is in matrix format. If TRUE, feature names are extracted from column names. Default is FALSE.}

\item{print_result}{Logical indicating whether to print the results of each manipulation step. Default is FALSE.}
}
\value{
A vector of filtered feature names after removing problematic variables.
}
\description{
This function performs feature manipulation on a given matrix or data frame by removing variables
with missing values, non-numeric variables, infinite values, and variables with zero standard deviation.
}
\examples{
# Load example data
data("eset_stad", package = "IOBR")
# Filter features with NA or outliers
feas <- feature_manipulation(data = eset_stad, feature = rownames(eset_stad), is_matrix = TRUE)
}
\author{
Dongqiang Zeng
}
