% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_select}
\alias{feature_select}
\title{Feature selection}
\usage{
feature_select(
  x,
  y,
  method = c("cor", "dif"),
  family = c("spearman", "pearson"),
  cutoff = NULL,
  padjcut = NULL
)
}
\arguments{
\item{x}{input matrix.Rownames should be features like gene symbols or cgi, colnames be samples}

\item{y}{response variable. Data type can be quantitative, binary and survival. Survival type can be generated through ?survival::Surv}

\item{method}{Binary for method = "dif", quantitative response value for "dif" and "cor".}

\item{family}{For method="cor", useser can choose "spearman" or "pearson" .}

\item{cutoff}{Numeric. Estimate and log2FC cutoff value for correlation analysis and limma dif analysis.}

\item{padjcut}{Numeric. Adjust P value cutoff.}
}
\value{
Returns a vector of selected feature names based on the specified criteria.
}
\description{
This function selects features based on correlation or differential expression methods. 
It integrates tools from the limma package for differential expression and basic statistical tests for correlation analysis.
}
\examples{

data("imvigor210_eset",package = "IOBR")
data("imvigor210_pdata", package = "IOBR")

mad <- apply(imvigor210_eset, 1, mad)
imvigor210_eset <- imvigor210_eset[mad > 0.5, ]
pd1 <- as.numeric(imvigor210_eset["PDCD1", ])
group <- ifelse(pd1 > mean(pd1), "high", "low")
pd1_cor <- feature_select(x = imvigor210_eset, y = pd1, method = "cor", family = "pearson", padjcut = 0.05, cutoff = 0.5)
pd1_dif <- feature_select(x = imvigor210_eset, y = pd1, method = "dif", padjcut = 0.05, cutoff = 2)
pd1_dif_2 <- feature_select(x = imvigor210_eset, y = group, method = "dif", padjcut = 0.05, cutoff = 2)
}
