% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers_in_bulk.R
\name{find_markers_in_bulk}
\alias{find_markers_in_bulk}
\title{Title}
\usage{
find_markers_in_bulk(
  pdata,
  eset,
  group,
  id_pdata = "ID",
  nfeatures = 200,
  top_n = 20,
  thresh.use = 0.25,
  only.pos = TRUE,
  min.pct = 0.25
)
}
\arguments{
\item{pdata}{}

\item{eset}{}

\item{group}{}

\item{id_pdata}{}

\item{nfeatures}{}

\item{top_n}{}

\item{thresh.use}{}

\item{only.pos}{}

\item{min.pct}{}
}
\description{
Title
}
