% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mutations.R
\name{find_mutations}
\alias{find_mutations}
\title{Mutation Analysis Related to Signature Scores}
\usage{
find_mutations(
  mutation_matrix,
  signature_matrix,
  id_signature_matrix = "ID",
  signature,
  min_mut_freq = 0.05,
  plot = TRUE,
  method = "multi",
  point.alpha = 0.1,
  save_path = NULL,
  palette = "jco",
  show_plot = TRUE,
  show_col = FALSE,
  width = 8,
  height = 4,
  oncoprint_group_by = "mean",
  oncoprint_col = "#224444",
  gene_counts = 10,
  jitter = FALSE,
  genes = NULL,
  point_size = 4.5
)
}
\arguments{
\item{mutation_matrix}{A matrix with sample names in the rows and genes in the columns representing mutation data.}

\item{signature_matrix}{A dataframe with identifier and target signatures.}

\item{id_signature_matrix}{The column name in the signature_matrix that acts as the identifier.}

\item{signature}{The name of the target signature for analysis.}

\item{min_mut_freq}{The minimum frequency of mutations required to consider a gene in the analysis.}

\item{plot}{A logical value; if TRUE, results will be plotted and saved at `save_path`.}

\item{method}{Specifies the statistical test method; "multi" for both Cuzick and Wilcoxon tests or "Wilcoxon" for only the Wilcoxon test.}

\item{save_path}{The directory path where plots and statistical analysis results will be saved.}

\item{palette}{The color palette for box plots.}

\item{show_plot}{A logical variable; if TRUE, plots will be rendered on display.}

\item{show_col}{Whether to show the color code of the palette.}

\item{width}{The width of the oncoprint plot.}

\item{height}{The height of the oncoprint plot.}

\item{oncoprint_group_by}{Specifies how to group the oncoprint, options are "mean" or "quantile".}

\item{oncoprint_col}{The color for mutations in the oncoprint.}

\item{gene_counts}{The number of genes to display in the oncoprint.}

\item{jitter}{A logical value; if TRUE, adds jitter to the box plot points.}

\item{genes}{A vector of gene names for drawing; if NULL, genes are selected based on mutation frequency.}

\item{point_size}{The size of points in the box plot.}

\item{point_alpha}{The transparency level of points in the box plot.}
}
\value{
Returns a list containing results from the statistical tests, oncoprint plots, and box plots.
}
\description{
This function searches for mutations related to a specific signature score and conducts statistical tests to evaluate
their significance. It can generate oncoprints and box plots to visualize the relationships and save the results.
}
\examples{
# Load mutation and signature data
maf_file <- "path_to_maf_file"
mut_list <- make_mut_matrix(maf = maf_file, isTCGA = TRUE, category = "multi")
mut <- mut_list$snp
# Assume 'tcga_stad_sig' is preloaded with appropriate data
results <- find_mutations(
  mutation_matrix = mut, signature_matrix = tcga_stad_sig,
  id_signature_matrix = "ID", signature = "CD_8_T_effector",
  min_mut_freq = 0.01, plot = TRUE, method = "multi",
  save_path = "path_to_save_results"
)
}
\author{
Dongqiang Zeng
}
