% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outlier_samples.R
\name{find_outlier_samples}
\alias{find_outlier_samples}
\title{Title}
\usage{
find_outlier_samples(
  eset,
  yinter = -3,
  project = "NSCLC",
  plot_hculst = FALSE,
  show_plot = TRUE
)
}
\arguments{
\item{eset}{expression set}

\item{yinter}{between -3 to -2}

\item{project}{default is NSCLC}

\item{show_plot}{}
}
\description{
Title
}
\examples{
使用WGCNA查看样本离异度
}
