% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_variable_genes.R, R/find_variable_genes2.R
\name{find_variable_genes}
\alias{find_variable_genes}
\title{Identify Variable Genes in Expression Data}
\usage{
find_variable_genes(
  eset,
  data_type = c("count", "normalized"),
  methods = c("low", "mad"),
  prop = 0.7,
  quantile = c(0.75, 0.5, 0.25),
  min.mad = 0.1,
  feas = NULL
)

find_variable_genes(
  eset,
  data_type = c("count", "normalized"),
  methods = c("low", "mad"),
  prop = 0.7,
  quantile = c(0.75, 0.5, 0.25),
  min.mad = 0.1,
  feas = NULL
)
}
\arguments{
\item{eset}{Numeric matrix. Gene expression data (genes as rows,
samples as columns).}

\item{data_type}{Character. Type of data: `"count"` or `"normalized"`.
Default is `"count"`.}

\item{methods}{Character vector. Methods for gene selection: `"low"`,
`"mad"`. Default is `c("low", "mad")`.}

\item{prop}{Numeric. Proportion of samples in which a gene must be expressed.
Default is 0.7.}

\item{quantile}{Numeric. Quantile threshold for minimum MAD (0.25, 0.5, 0.75).
Default is 0.75.}

\item{min.mad}{Numeric. Minimum allowable MAD value. Default is 0.1.}

\item{feas}{Character vector or `NULL`. Additional features to include.
Default is `NULL`.}
}
\value{
Matrix subset of `eset` containing variable genes.

Matrix subset of `eset` containing variable genes.
}
\description{
Identifies variable genes from a gene expression dataset using specified
selection criteria. Supports multiple methods, including expression
thresholding and variability estimation via median absolute deviation (MAD).

Identifies variable genes from a gene expression dataset using specified
selection criteria. Supports multiple methods, including expression
thresholding and variability estimation via median absolute deviation (MAD).
}
\examples{
\donttest{
eset_tme_stad <- load_data("eset_tme_stad")
eset <- find_variable_genes(
  eset = eset_tme_stad,
  data_type = "normalized",
  methods = "mad",
  quantile = 0.25
)
}
\donttest{
eset_tme_stad <- load_data("eset_tme_stad")
eset <- find_variable_genes(
  eset = eset_tme_stad,
  data_type = "normalized",
  methods = "mad",
  quantile = 0.25
)
}
}
\author{
Dongqiang Zeng
}
