% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_variable_genes2.R
\name{find_variable_genes}
\alias{find_variable_genes}
\title{Title}
\usage{
find_variable_genes(
  eset,
  data_type = c("count", "normalized"),
  methods = c("low", "mad"),
  prop = 0.7,
  quantile = c(0.75, 0.5, 0.25),
  min.mad = 0.1,
  feas = NULL
)
}
\arguments{
\item{feas}{}
}
\description{
Title
}
