% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_variable_genes.R
\name{find_variable_genes}
\alias{find_variable_genes}
\title{Title find_variable_genes}
\usage{
find_variable_genes(
  x,
  data_type = c("count", "normalized"),
  methods = c("low", "mad"),
  prop = 0.7,
  quantile = c(0.75, 0.5, 0.25),
  min.mad = 0.1,
  feas = NULL
)
}
\arguments{
\item{x}{A gene expression matrix. It should be a numeric matrix or a matrix-like object.}

\item{data_type}{The type of data in the gene expression matrix. The default value is "count", indicating count data. This parameter determines the method used for filtering variable genes.}

\item{methods}{A character vector specifying the filtering methods to be applied. The default value is c("low", "mad"), indicating both low expression filtering and median absolute deviation (MAD) filtering methods.}

\item{prop}{The proportion threshold for low expression filtering. It specifies the minimum proportion of samples in which a gene should be expressed to be considered as a variable gene. The default value is 0.7, meaning genes expressed in more than 7% of the samples will be kept.}

\item{quantile}{The quantile value for MAD filtering. It determines the threshold for selecting genes based on their median absolute deviation. The default value is 0.75, indicating the top 75% of genes with higher MAD will be selected.}

\item{min.mad}{The minimum median absolute deviation threshold for MAD filtering. It specifies the minimum MAD value a gene should have to be considered as a variable gene. The default value is 0.1.}

\item{feas}{An optional vector of gene names or indices specifying additional genes to be considered as variable genes. This allows users to manually include certain genes in the filtered matrix. The default value is NULL.}
}
\description{
The "find_variable_genes" function is designed to identify and filter variable genes in a gene expression matrix based on the specified criteria. It takes a gene expression matrix ("x") as input and returns a filtered matrix ("fx") containing only the variable genes.
}
\author{
Dongqiang Zeng
}
