% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_msigdb.R
\name{format_msigdb}
\alias{format_msigdb}
\title{Format Input Signatures from MSigDB}
\usage{
format_msigdb(gmt, ont = "term", gene = "gene")
}
\arguments{
\item{gmt}{Character string. Path to a GMT file.}

\item{ont}{Character string. Name of the signature/set column in the parsed
GMT table. Default is `"term"`.}

\item{gene}{Character string. Name of the gene column in the parsed GMT
table. Default is `"gene"`.}
}
\value{
Named list of character vectors, where each element contains the
  genes belonging to one signature.
}
\description{
Reads a GMT file from MSigDB and converts it into a named list of gene sets
suitable for IOBR functions.
}
\examples{
\donttest{
tf <- tempfile(fileext = ".gmt")
writeLines(
  c(
    "HALLMARK_TNFA_SIGNALING_VIA_NFKB\tNA\tTNF\tNFKB1\tNFKB2",
    "HALLMARK_P53_PATHWAY\tNA\tTP53\tMDM2\tCDKN1A"
  ),
  con = tf
)

sig_list <- format_msigdb(tf, ont = "term", gene = "gene")
names(sig_list)
sig_list[[1]]
}
}
