% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_msigdb.R
\name{format_msigdb}
\alias{format_msigdb}
\title{Format input signatures from MsgiDB}
\usage{
format_msigdb(gmt, ont = "term", gene = "gene")
}
\arguments{
\item{gmt}{`signature` data frame downloaded from `https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp#H`}

\item{ont}{column name of signatures}

\item{gene}{column name of gene}
}
\value{
signature list
}
\description{
Format input signatures from MsgiDB
}
\examples{
# Download GMT file from MSigDB:
# https://www.gsea-msigdb.org/gsea/msigdb/
format_msigdb(
  "E:/0-0-0-My Pipeline/00-Signature-collection/h.all.v2023.2.Hs.symbols.gmt",
  ont = "term"
)

}
