% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_signatures.R
\name{format_signatures}
\alias{format_signatures}
\title{Transform signature gene data into list format}
\usage{
format_signatures(sig_data, save_signature = FALSE, output_name = "signatures")
}
\arguments{
\item{sig_data}{signature data with signature name as column name and genes in row, using `NA` to substitude the missing value}

\item{save_signature}{logical, save siganture list as RData}

\item{output_name}{string of name to save signature RData}
}
\value{
list of signatures
}
\description{
Transform signature gene data into list format
}
\examples{

# load signature data frame
data("sig_excel", package = "IOBR")
# transform into gene list for IOBR functions
format_signatures(sig_excel)
reading each column and transfer it to a list
standerdized the name of list
}
\author{
Dongqiang Zeng
}
