% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_signatures.R
\name{format_signatures}
\alias{format_signatures}
\title{Transform Signature Data into List Format}
\usage{
format_signatures(sig_data, save_signature = FALSE, output_name = "signatures")
}
\arguments{
\item{sig_data}{Data frame with signature names as columns and genes in rows; use NA for missing values.}

\item{save_signature}{Logical indicating whether to save the signature list as RData. Default is FALSE.}

\item{output_name}{String for the output RData file name. Default is "signatures".}
}
\value{
A list of signatures.
}
\description{
This function converts signature data from a data frame (with signatures as columns and genes as rows)
into a list format suitable for IOBR functions, replacing NA values appropriately.
}
\examples{
# Load signature data frame
data("sig_excel", package = "IOBR")
# Transform into gene list for IOBR functions
format_signatures(sig_excel)
reading each column and transfer it to a list
standerdized the name of list
}
\author{
Dongqiang Zeng
}
