% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRef.R
\name{generateRef}
\alias{generateRef}
\title{Title}
\usage{
generateRef(dat, pheno, mode = c("oneVSothers", "pairs"), FDR)
}
\arguments{
\item{dat}{data frame or matrix; gene probes in the row and samples in the column}

\item{pheno}{character vector; cell type class of the samples}

\item{mode}{"oneVSothers" or "pairs"; two modes for identifying significantly differentially expressed genes}

\item{FDR}{numeric; genes with BH adjust p value < FDR are considered significant.}
}
\value{

}
\description{
Title
}
\examples{
data("Stromal_v2.pheno")
data("Stromal_v2")
immune_feature <- generateRef(dat = Stromal_v2, pheno = Stromal_v2.pheno, mode = "oneVSothers", FDR = 0.05)
reference <- immune_feature$reference_matrix
condition_number <- immune_feature$condition_number
top_probe <- immune_feature$G
}
