% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRef.R
\name{generateRef}
\alias{generateRef}
\title{Generate Reference signature matrix}
\usage{
generateRef(dds, pheno, FDR = 0.05, dat, method = "limma")
}
\arguments{
\item{dds}{raw count data from RNA-seq; Necessary if used the method DESeq2}

\item{pheno}{character vector; cell type class of the samples}

\item{FDR}{numeric; genes with BH adjust p value < FDR are considered significant.}

\item{dat}{data frame or matrix; normalized transcript quantification data (like FPKM, TPM). Note: cell's median expression level of the identified probes will be the output of reference_matrix.}

\item{method}{limma or DESeq2}
}
\description{
Generate Reference signature matrix
}
