% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRef.R
\name{generateRef}
\alias{generateRef}
\title{Generate reference gene matrix}
\usage{
generateRef(dat, pheno, mode = c("oneVSothers", "pairs"), FDR = 0.05)
}
\arguments{
\item{dat}{data frame or matrix; gene probes in the row and samples in the column}

\item{pheno}{character vector; cell type class of the samples}

\item{mode}{"oneVSothers" or "pairs"; two modes for identifying significantly differential expressed genes}

\item{FDR}{numeric; genes with BH adjust p value < FDR are considered significant.}
}
\value{

}
\description{
Generate reference gene matrix
}
\author{
Rongfang Shen
}
