% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRef_DEseq2.R
\name{generateRef_DEseq2}
\alias{generateRef_DEseq2}
\title{Generate Reference signature matrix using DEseq2}
\usage{
generateRef_DEseq2(dds, pheno, FDR = 0.05, dat)
}
\arguments{
\item{dds}{raw count data from RNA-seq}

\item{pheno}{character vector; cell type class of the samples}

\item{FDR}{numeric; genes with BH adjust p value < FDR are considered significant.}

\item{dat}{data frame or matrix; normalized transcript quantification data (like FPKM, TPM). Note: cell's median expression level of the identified probes will be the output of reference_matrix.}
}
\description{
Generate Reference signature matrix using DEseq2
}
