% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRef_limma.R
\name{generateRef_limma}
\alias{generateRef_limma}
\title{Generate Reference signature matrix using limma}
\usage{
generateRef_limma(dat, pheno, FDR = 0.05)
}
\arguments{
\item{dat}{data frame or matrix; gene probes in the row and samples in the column}

\item{pheno}{character vector; cell type class of the samples}

\item{FDR}{numeric; genes with BH adjust p value < FDR are considered significant.}
}
\description{
Generate Reference signature matrix using limma
}
