% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRef_rnaseq.R
\name{generateRef_rnaseq}
\alias{generateRef_rnaseq}
\title{Generate reference gene matrix from DEGs result}
\usage{
generateRef_rnaseq(dds, pheno, mode = "oneVSothers", FDR = 0.05, dat)
}
\arguments{
\item{dds}{raw count data from RNA-seq}

\item{pheno}{data frame or matrix;un-normalized counts of RNA-seq}

\item{mode}{mode "oneVSothers" or "pairs"; two modes for identifying significantly differentially expressed genes}

\item{FDR}{character vector; cell type class of the samples}

\item{dat}{data frame or matrix; normalized transcript quantification data (like FPKM, TPM). Note: cell's median expression level of the identified probes will be the output of reference_matrix.}
}
\value{

}
\description{
Generate reference gene matrix from DEGs result
}
\author{
Rongfang Shen
}
