% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols.R
\name{get_cols}
\alias{get_cols}
\title{Set and view the color palettes}
\usage{
get_cols(cols = "normal", palette = 1, show_col = T, seed = 123)
}
\arguments{
\item{cols}{Vector of colors, users can define the cols manually.  This may also be a single character, such as normal and random,  to a palette as specified by `palettes(){IOBR}`
See [palettes](http://127.0.0.1:60491/help/library/IOBR/html/palettes.html) for details}

\item{palette}{Numeric value corresponding with color palette. Default is 1, other options: 2, 3, 4}

\item{show_col}{Whether to show color palettes}

\item{seed}{Seed of the random number generator, default is 123. The parameter works when cols ="random"}
}
\value{
Vector of colors
}
\description{
Set and view the color palettes
}
\examples{
mycols<-get_cols()
}
