% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor.R
\name{get_cor}
\alias{get_cor}
\title{Calculate and Visualize Correlation between Two Variables}
\usage{
get_cor(
  eset,
  pdata = NULL,
  id_pdata = "ID",
  var1,
  var2,
  scale = TRUE,
  subtype = NULL,
  na.subtype.rm = FALSE,
  color_subtype = NULL,
  palette = "jama",
  index = NULL,
  method = "spearman",
  show_cor_result = T,
  col_line = NULL,
  id = "NULL",
  show_lebel = FALSE,
  point_size = 4,
  title = NULL,
  alpha = 0.7,
  title_size = 2,
  text_size = 15,
  axis_angle = 0,
  hjust = 0,
  show_plot = T,
  fig.format = "png",
  fig.width = 7,
  fig.height = 7.3,
  path = NULL,
  save_plot = FALSE
)
}
\arguments{
\item{eset}{A dataset containing the variables of interest.}

\item{pdata}{An optional data frame providing additional data for the analysis. Default is NULL.}

\item{id_pdata}{The column name in pdata that contains unique identifiers. Default is "ID".}

\item{var1}{The name of the first variable to analyze.}

\item{var2}{The name of the second variable to analyze.}

\item{scale}{Whether to scale the data. Default is TRUE.}

\item{subtype}{An optional variable that represents subtypes/categories in the analysis. Default is NULL.}

\item{na.subtype.rm}{Whether to remove NA values in the subtype variable. Default is FALSE.}

\item{color_subtype}{The color palette for the subtypes/categories. Default is NULL.}

\item{palette}{The name of the color palette. Default is "jama".}

\item{index}{The index of the plot. Default is NULL.}

\item{method}{The correlation method to use. Default is "spearman".}

\item{show_cor_result}{Whether to print the correlation result. Default is TRUE.}

\item{col_line}{The color of the line in the correlation plot. Default is NULL.}

\item{id}{The ID column name to use for labeling points on the plot. Default is "NULL".}

\item{show_lebel}{Whether to show labels on the plot. Default is FALSE.}

\item{point_size}{The size of the points on the plot. Default is 4.}

\item{title}{The title of the plot. Default is NULL.}

\item{alpha}{The transparency level of points on the plot. Default is 0.7.}

\item{title_size}{The size of the title on the plot. Default is 2.}

\item{text_size}{The size of the text on the plot. Default is 15.}

\item{axis_angle}{The rotation angle of the axes labels on the plot. Default is 0.}

\item{hjust}{The horizontal justification of the axes labels on the plot. Default is 0.}

\item{show_plot}{Whether to display the plot. Default is TRUE.}

\item{fig.format}{The format of the saved figure. Default is "png".}

\item{fig.width}{The width of the saved figure. Default is 7.}

\item{fig.height}{The height of the saved figure. Default is 7.3.}

\item{path}{The path where the figure will be saved. Default is NULL.}

\item{save_plot}{Whether to save the plot as a file. Default is FALSE.}
}
\description{
The "get_cor" function calculates and visualizes the correlation between two variables in a dataset. It provides options to scale the data, handle missing values, and incorporate additional data. The function supports various correlation methods and can display the correlation result. It generates a correlation plot with optional subtypes or categories, including a regression line. The plot can be customized with color palettes, labels, and titles. Additionally, the function allows saving the plot and data for further analysis.
}
\author{
Dongqiang Zeng
}
