% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_var_fea.R
\name{high_var_fea}
\alias{high_var_fea}
\title{Extract top varity features form statistical results}
\usage{
high_var_fea(
  result,
  target,
  name_padj = "padj",
  padj_cutoff = 1,
  name_logfc,
  logfc_cutoff = 0,
  n = 10
)
}
\arguments{
\item{result}{a tibble}

\item{target}{name of variables}
}
\value{

}
\description{
Extract top varity features form statistical results
}
\author{
Dongqiang Zeng
}
