% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobr_cor_plot.R
\name{iobr_cor_plot}
\alias{iobr_cor_plot}
\title{Integrative correlation between phenotype and features}
\usage{
iobr_cor_plot(
  pdata_group,
  id1 = "ID",
  feature_matrix,
  id2 = "ID",
  target,
  group = "group3",
  is_target_continuous = TRUE,
  padj_cutoff = 1,
  index = 1,
  panel = panel_for_signature,
  category = "signature",
  signature_groups = signature_group,
  ProjectID = "TCGA-STAD",
  palette_box = "nrc",
  palette_corplot = "pheatmap",
  palette_heatmap = 2,
  feature_limit = 26,
  character_limit = 30,
  show_heatmap_col_name = FALSE,
  show_col = FALSE
)
}
\arguments{
\item{pdata_group}{pdata}

\item{id1}{column of identifier of pdata}

\item{feature_matrix}{feature matrix}

\item{id2}{column of identifier of feature matrix}

\item{target}{column of target variables}

\item{group}{column used to differentiate patient groups}

\item{is_target_continuous}{logical variable, if TRUE, new group will be generated based on the average or the third percentile}

\item{index}{index use to order the file name}

\item{panel}{panel for signature group `signature_group` or signature gene `gene_group`}

\item{category}{`signature` or `gene`}

\item{show_heatmap_col_name}{logical variable, if TRUE, tidyheatmap will print the column name}

\item{show_col}{logical variable, if TRUE, color will be print and show in the R studio}
}
\value{

}
\description{
Integrative correlation between phenotype and features
}
\examples{

}
\author{
Dongqiang Zeng
}
