% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobr_deg.R
\name{iobr_deg}
\alias{iobr_deg}
\title{Differential expressed analyses}
\usage{
iobr_deg(
  eset,
  annoation = annotation_grch38,
  pdata,
  group_id = "group",
  pdata_id = "ID",
  array = TRUE,
  method = "DESeq2",
  contrast = c("deg_group", "High", "Low"),
  path = NULL,
  padj_cutoff = 0.01,
  logfc_cutoff = 0.5,
  volcano_plot = FALSE,
  col_volcano = 1,
  heatmap = TRUE,
  col_heatmap = 1
)
}
\arguments{
\item{eset}{Expression set object containing the matrix of gene expression data. Default is NULL.}

\item{annoation}{The annotation to be used. Default is annotation_grch38.}

\item{pdata}{The data frame containing the sample information.}

\item{group_id}{Specifies the ID of the group. Default is "group".}

\item{pdata_id}{Specifies the ID of the pdata. Default is "ID".}

\item{array}{Logical value to specify whether the 'eset' is a microarray data or not. Default is TRUE.}

\item{method}{The method used for DEG analysis: either "DESeq2" or "limma". Default is "DESeq2".}

\item{contrast}{Specifies groups to apply DESeq2 results() function, it is typically applied to groups of interest in terms of an outcome variable of interest. Default is c("deg_group","High","Low").}

\item{path}{}

\item{padj_cutoff}{Specifies the cutoff of adjusted p-values. Default is 0.01.}

\item{logfc_cutoff}{Specifies the cutoff of log2 Fold Change. Default is 0.5.}

\item{volcano_plot}{Logical value to specify whether to plot volcano plots or not. Default is FALSE.}

\item{col_volcano}{Specifies the color of the volcano plots. Default is 1.}

\item{heatmap}{Logical value to specify whether to plot heatmaps or not. Default is TRUE.}

\item{col_heatmap}{Specifies the color of the heatmaps. Default is 1.}
}
\value{
DEG object
}
\description{
Differential expressed analyses
}
\author{
Dongqiang Zeng
}
