% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobr_deg.R
\name{iobr_deg}
\alias{iobr_deg}
\title{iobr_deg}
\usage{
iobr_deg(
  eset,
  annotation = NULL,
  pdata,
  group_id = "group",
  pdata_id = "ID",
  array = FALSE,
  method = "DESeq2",
  contrast = c("High", "Low"),
  path = NULL,
  padj_cutoff = 0.01,
  logfc_cutoff = 0.5,
  volcano_plot = FALSE,
  col_volcano = 1,
  heatmap = TRUE,
  col_heatmap = 1,
  parallel = FALSE
)
}
\arguments{
\item{eset}{The gene expression data matrix}

\item{annotation}{(Optional) The annotation object for mapping gene IDs to gene names. Default is NULL.}

\item{pdata}{The DataFrame containing sample information and group labels.}

\item{group_id}{The column name in "pdata" that specifies the group labels. Default value is "group".}

\item{pdata_id}{The column name in "pdata" that specifies the sample IDs. Default value is "ID".}

\item{array}{Logical value indicating whether to normalize the gene expression data using quantile normalization. Default value is TRUE.}

\item{method}{The method to be used for differential expression analysis. Options are "DESeq2" or "limma". Default value is "DESeq2".}

\item{contrast}{A character vector specifying the contrast for the analysis.  The first and second elements are the names of the two groups being compared. Default value is  "High" and "Low".}

\item{path}{path to save results}

\item{padj_cutoff}{The adjusted p-value cutoff for determining significance. Default value is 0.01.}

\item{logfc_cutoff}{The log2 fold change cutoff for determining significance. Default value is 0.5.}

\item{volcano_plot}{Logical value indicating whether to generate a volcano plot. Default value is FALSE.}

\item{col_volcano}{The color index for plotting the volcano plot. Default value is 1.}

\item{heatmap}{Logical value indicating whether to generate a heatmap. Default value is TRUE.}

\item{col_heatmap}{The color index for plotting the heatmap. Default value is 1.}

\item{parallel}{default is FALSE}
}
\value{
DEG object
}
\description{
This function performs differential expression analysis on gene expression data using the DESeq2 or limma method. It filters low count data, calculates fold changes and adjusted p-values, and identifies differentially expressed genes (DEGs) based on specified cutoffs. It also provides optional visualization tools such as volcano plots and heatmaps. The function saves the results as an RData file and an Excel file.
}
\author{
Dongqiang Zeng
}
