% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobr_pca.R
\name{iobr_pca}
\alias{iobr_pca}
\title{Title}
\usage{
iobr_pca(
  data,
  is.matrix = TRUE,
  scale = TRUE,
  is.log = FALSE,
  pdata,
  id_pdata = "ID",
  group = NULL,
  geom.ind = "point",
  cols = "normal",
  palette = "jama",
  repel = FALSE,
  ncp = 5,
  axes = c(1, 2),
  addEllipses = TRUE
)
}
\arguments{
\item{geom.ind}{}
}
\description{
Title
}
\author{
Dongqiang Zeng
}
