% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobr_pca.R
\name{iobr_pca}
\alias{iobr_pca}
\title{Principal Component Analysis (PCA) Visualization Function}
\usage{
iobr_pca(
  data,
  is.matrix = TRUE,
  scale = TRUE,
  is.log = FALSE,
  pdata,
  id_pdata = "ID",
  group = NULL,
  geom.ind = "point",
  cols = "normal",
  palette = "jama",
  repel = FALSE,
  ncp = 5,
  axes = c(1, 2),
  addEllipses = TRUE
)
}
\arguments{
\item{data}{The input data for PCA. It should be a matrix or a data frame.}

\item{is.matrix}{Specifies whether the input data is a matrix. Default is TRUE.}

\item{scale}{Specifies whether to scale the input data. Default is TRUE.}

\item{is.log}{Specifies whether to log transform the input data. Default is FALSE.}

\item{pdata}{Additional data associated with the principal components. It should be a data frame. Default is NULL.}

\item{id_pdata}{The column name in 'pdata' that represents the ID for matching with 'data'. Default is "ID".}

\item{group}{The column name in 'pdata' that represents groups/categories to color the points. Default is NULL.}

\item{geom.ind}{The type of geometric representation for the points in the PCA plot. Default is "point".}

\item{cols}{The color scheme to be used for group categories. Default is "normal".}

\item{palette}{The color palette to be used for group categories. Default is "jama".}

\item{repel}{Specifies whether to repel the data points to avoid overlap. Default is FALSE.}

\item{ncp}{The number of dimensions to keep in the PCA. Default is 5.}

\item{axes}{The dimensions/axes to be plotted. Default is c(1, 2).}

\item{addEllipses}{Specifies whether to add concentration ellipses to the plot. Default is TRUE.}
}
\description{
The iobr_pca function performs Principal Component Analysis (PCA), which reduces the dimensionality of data while maintaining most of the original variance, and visualizes the PCA results on a scatter plot.
}
\examples{
data("eset_stad", package = "IOBR")
eset <- count2tpm(eset_stad)
iobr_pca(eset, is.matrix = TRUE, scale = TRUE, is.log = TRUE, pdata = stad_group, id_pdata = "ID", group = "subtype")

}
\author{
Dongqiang Zeng
}
