% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobr_pca.R
\name{iobr_pca}
\alias{iobr_pca}
\title{Principal Component Analysis (PCA) Visualization}
\usage{
iobr_pca(
  data,
  is.matrix = TRUE,
  scale = TRUE,
  is.log = FALSE,
  pdata,
  id_pdata = "ID",
  group = NULL,
  geom.ind = "point",
  cols = "normal",
  palette = "jama",
  repel = FALSE,
  ncp = 5,
  axes = c(1, 2),
  addEllipses = TRUE
)
}
\arguments{
\item{data}{Input data for PCA: matrix or data frame.}

\item{is.matrix}{Logical indicating if input is a matrix. Default is TRUE.}

\item{scale}{Logical indicating whether to scale the data. Default is TRUE.}

\item{is.log}{Logical indicating whether to log-transform the data. Default is FALSE.}

\item{pdata}{Data frame with sample IDs and grouping information.}

\item{id_pdata}{Column name in `pdata` for sample IDs. Default is "ID".}

\item{group}{Column name in `pdata` for grouping variable. Default is NULL.}

\item{geom.ind}{Type of geometric representation for points. Default is "point".}

\item{cols}{Color scheme for groups. Default is "normal".}

\item{palette}{Color palette for groups. Default is "jama".}

\item{repel}{Logical indicating whether to repel overlapping points. Default is FALSE.}

\item{ncp}{Number of principal components to retain. Default is 5.}

\item{axes}{Principal components to plot (e.g., c(1, 2)). Default is c(1, 2).}

\item{addEllipses}{Logical indicating whether to add concentration ellipses. Default is TRUE.}
}
\value{
A ggplot object of the PCA plot.
}
\description{
This function performs Principal Component Analysis (PCA) on gene expression data,
reduces dimensionality while preserving variance, and generates a scatter plot visualization.
}
\examples{
data("eset_stad", package = "IOBR")
eset <- count2tpm(eset_stad)
iobr_pca(eset, is.matrix = TRUE, scale = TRUE, is.log = TRUE, pdata = stad_group, id_pdata = "ID", group = "subtype")

}
\author{
Dongqiang Zeng
}
