% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_calculation.R
\name{IPS_calculation}
\alias{IPS_calculation}
\title{This R-script can be used to calculate Immunophenoscore (IPS)}
\usage{
IPS_calculation(project, eset, plot)
}
\arguments{
\item{eset}{expression data from tab-delimited text file, with official human gene symbols (HGNC) in the rowname;
expression values (i.e. log2(TPM+1) for each sample in columns}

\item{plot}{default = FALSE;Needs packages ggplot2,grid,gridExtra}
}
\value{
IPS score data frame
}
\description{
This R-script can be used to calculate Immunophenoscore (IPS)
}
\examples{

鍋囧瑕佺敾鍥剧殑璇㤼㹤
file to deposit resluts
棣栧厛鏋勫缓鍥剧殑鍌ㄥ瓨鏂囦欢澶戼㸹
}
