% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_calculation.R
\name{IPS_calculation}
\alias{IPS_calculation}
\title{Title This R-script can be used to calculate Immunophenoscore (IPS)}
\usage{
IPS_calculation(project, eset, plot)
}
\arguments{
\item{eset}{expression data from tab-delimited text file, with official human gene symbols (HGNC) in the first columns;
expression values (i.e. log2(TPM+1) for each sample in the other columns}

\item{plot}{default = FALSE;Needs packages ggplot2,grid,gridExtra}
}
\value{
IPS score data frame
}
\description{
Title This R-script can be used to calculate Immunophenoscore (IPS)
}
\examples{
ips_crc <- IPS_calculation(project = "TCGA-COAD", eset = eset_crc, plot = FALSE)
假如要画图的话
file to deposit resluts
首先构建图的储存文件夹
}
