% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_calculation.R
\name{IPS_calculation}
\alias{IPS_calculation}
\title{This R-script can be used to calculate Immunophenoscore (IPS)}
\usage{
IPS_calculation(project = NULL, eset, plot = FALSE)
}
\arguments{
\item{project}{default is NULL}

\item{eset}{expression data from tab-delimited text file, with official human gene symbols (HGNC) in the rowname;
expression values (i.e. log2(TPM+1) for each sample in columns}

\item{plot}{default = FALSE;Needs packages ggplot2, grid, gridExtra}
}
\value{
IPS score data frame
}
\description{
This R-script can be used to calculate Immunophenoscore (IPS)
}
\examples{
file to deposit reslut
首先构建图的储存文件夹
}
