% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_helper.R
\name{ipsmap}
\alias{ipsmap}
\title{Map Score to Immunophenoscore}
\usage{
ipsmap(x)
}
\arguments{
\item{x}{Numeric value representing the aggregate z-score.}
}
\value{
Integer value between 0 and 10 representing the Immunophenoscore.
}
\description{
Maps input score to Immunophenoscore (IPS) on a 0-10 scale. Scores \eqn{\le}0 map to 0,
scores \eqn{\ge}3 map to 10, and intermediate scores are linearly scaled.
}
\examples{
ips <- ipsmap(2.5)
ips <- ipsmap(-1)
ips <- ipsmap(5)
}
