% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{limma.dif}
\alias{limma.dif}
\title{limma.dif}
\usage{
limma.dif(exprdata, pdata, contrastfml)
}
\arguments{
\item{exprdata}{input matrix.Rownames should be features like gene symbols or cgi, colnames be samples}

\item{pdata}{phenotype data.Two-column dataframe which column 1 should be the same with the colnames of exprdata and column 2 are the grouping variable.}

\item{contrastfml}{see ?makeContrasts}
}
\value{
a data frame return by limma::toptable, which genes at rows and following columns: genelist, logFC, AveExpr...
}
\description{
This function performs differential expression analysis using the limma package on a given gene expression dataset. 
It constructs a design matrix from phenotype data, fits a linear model, applies contrasts, and computes statistics for 
differential expression.
}
