% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2eset.R
\name{log2eset}
\alias{log2eset}
\title{Judge whether the expression set needs log2 transformation}
\usage{
log2eset(eset = NULL)
}
\arguments{
\item{eset}{gene expression set with row as genes and sample in the column}
}
\value{
gene expression set
}
\description{
Judge whether the expression set needs log2 transformation
}
