% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_cal.R
\name{LR_cal}
\alias{LR_cal}
\title{LR_cal}
\usage{
LR_cal(
  eset,
  data_type = c("count", "tpm"),
  id_type = "ensembl",
  cancer_type = "pancan"
)
}
\arguments{
\item{eset}{ExpressionSet object containing gene expression data.}

\item{data_type}{Type of input data, possible values are "count" or "tpm".}

\item{id_type}{Type of gene identifier, default is "ensembl".}
}
\value{
A data frame containing ligand-receptor interaction scores
}
\description{
Quantifies ligand-receptor interactions in the tumor microenvironment from bulk gene expression
}
\examples{
data("eset_stad", package = "IOBR")
lr <- LR_cal(eset = eset_stad, data_type = "count", id_type = "ensembl")
}
\references{
Lapuente-Santana, van Genderen, M., Hilbers, P., Finotello, F., & Eduati, F. (2021). 'Interpretable systems biomarkers predict response to immune-checkpoint inhibitors.' Patterns (New York, N.Y.), 2(8), 100293. https://doi.org/10.1016/j.patter.2021.100293
}
