% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_duplicate.R
\name{merge_duplicate}
\alias{merge_duplicate}
\title{Merge data frames with duplicated column names}
\usage{
merge_duplicate(x, y, by.x, by.y, all.x, all.y, all = NULL, choose = "x")
}
\arguments{
\item{choose}{}
}
\description{
Merge data frames with duplicated column names
}
