% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_eset.R
\name{merge_eset}
\alias{merge_eset}
\title{Merge expression sets by row names}
\usage{
merge_eset(eset1, eset2, eset3 = NULL)
}
\arguments{
\item{eset1}{eset1}

\item{eset2}{eset2}

\item{eset3}{eset3, default is NULL}
}
\description{
Merge expression sets by row names
}
\author{
Dongqiang Zeng
}
