% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{microenvironmentScores}
\alias{microenvironmentScores}
\title{Calculate microenvironment scores}
\usage{
microenvironmentScores(adjustedScores)
}
\arguments{
\item{adjustedScores}{the combined microenvironment scores}
}
\value{
the microenvironment scores
}
\description{
\code{microenvironmentScores} Returns the adjusted xCell scores.
}
