% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse2human_eset.R
\name{mouse2human_eset}
\alias{mouse2human_eset}
\title{Converting mouse gene symbol to human gene symbol of expression set.}
\usage{
mouse2human_eset(
  eset,
  source = "ensembl",
  is_matrix = T,
  column_of_symbol = NULL,
  verbose = FALSE
)
}
\arguments{
\item{eset}{expression matrix or data frame}

\item{source}{default is ensembl, if an error is reported, set parameter to `local` and mus_human_gene_symbol will be use to convert gene symbols}

\item{is_matrix}{default is T, if not, column_of_symbol must be se}

\item{column_of_symbol}{default is null}
}
\description{
Converting mouse gene symbol to human gene symbol of expression set.
}
