% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_sig.R
\name{output_sig}
\alias{output_sig}
\title{save signature data into a data frame}
\usage{
output_sig(signatures, format = "csv", file.name)
}
\arguments{
\item{signatures}{output signatures, a list or a string}

\item{format}{output format}

\item{file.name}{file name of output data}
}
\description{
save signature data into a data frame
}
