% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palettes}
\alias{palettes}
\title{Choose colors.}
\usage{
palettes(
  category = "box",
  palette = "nrc",
  alpha = 1,
  counts = 50,
  show_col = TRUE,
  show_message = FALSE
)
}
\arguments{
\item{category}{choose what type of plot you want to color,There are seven categories you can choose: `box`, `continue2`, `continue`, `random`, `heatmap`, `heatmap3`, `tidyheatmap`}

\item{palette}{Each category has some options to choose}

\item{alpha}{intensity of the color}

\item{counts}{number of colors, only affect continue colors}

\item{show_col}{logical variable, if TRUE, color will be print and show in the R studio}

\item{show_message}{print palettes that can be choose}
}
\value{
A vector of colors based on the selected options.
}
\description{
This function selects color palettes based on the type of plot you want to color. It supports multiple categories
and palette options, allowing for the adjustment of color intensity and the number of colors.
}
\examples{
# Get a palette for box plots with the 'nrc' palette
colors <- palettes(category = "box", palette = "nrc", show_message = TRUE)

# Get a continuous color scale for heatmaps
heatmap_colors <- palettes(category = "heatmap", palette = 1, counts = 100, show_col = TRUE)
}
\author{
Dongqiang Zeng
}
