% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_names.R
\docType{data}
\name{patterns_to_na}
\alias{patterns_to_na}
\title{Default pattern list for name cleaning}
\format{
A character vector of length 12.
}
\usage{
patterns_to_na
}
\description{
A character vector of common substrings to remove from feature names
(used in \code{remove_names()} and other helper functions).
}
\keyword{datasets}
