% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_bar.R
\name{percent_bar_plot}
\alias{percent_bar_plot}
\title{percent_bar_plot}
\usage{
percent_bar_plot(
  input,
  x,
  y,
  subset.x = NULL,
  color = NULL,
  palette = NULL,
  title = NULL,
  axis_angle = 0,
  add_Freq = TRUE,
  Freq = NULL,
  size_freq = 8,
  legend.size = 0.5,
  legend.size.text = 10,
  add_sum = T,
  print_result = T,
  round.num = 2
)
}
\arguments{
\item{input}{data frame}

\item{x}{group in x axis}

\item{y}{group in y axis}

\item{color}{colors}

\item{palette}{palette for box}

\item{title}{title}

\item{axis_angle}{range from [0,90]}

\item{add_Freq}{default is true for data frame}

\item{Freq}{column name of Frequency}

\item{add_sum}{default is TRUE}

\item{round.num}{}
}
\description{
percent_bar_plot
}
