% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_bar.R
\name{percent_bar_plot}
\alias{percent_bar_plot}
\title{Create a Percent Bar Plot}
\usage{
percent_bar_plot(
  input,
  x,
  y,
  subset.x = NULL,
  color = NULL,
  palette = NULL,
  title = NULL,
  axis_angle = 0,
  coord_flip = FALSE,
  add_Freq = TRUE,
  Freq = NULL,
  size_freq = 8,
  legend.size = 0.5,
  legend.size.text = 10,
  add_sum = T,
  print_result = T,
  round.num = 2
)
}
\arguments{
\item{input}{Input data frame.}

\item{x}{Name of the x-axis variable.}

\item{y}{Name of the y-axis variable.}

\item{subset.x}{Optional subset of x-axis values.}

\item{color}{Optional color palette for the bars.}

\item{palette}{Optional palette type for color selection.}

\item{title}{Optional title for the plot.}

\item{axis_angle}{Optional angle for the axis labels. range from [0,90]}

\item{coord_flip}{Boolean indicating whether to flip the x and y axes.}

\item{add_Freq}{default is true for data frame}

\item{Freq}{Optional name for the frequency column.}

\item{size_freq}{Size of the frequency labels.}

\item{legend.size}{Size of the legend.}

\item{legend.size.text}{Size of the legend text.}

\item{add_sum}{Boolean indicating whether to add the sum to the x-axis labels.}

\item{print_result}{Boolean indicating whether to print the result data frame.}

\item{round.num}{Number of decimal places to round the proportion column.}
}
\value{
A ggplot object representing the percentage bar plot.
}
\description{
This function generates a bar plot that visualizes the percentage distribution of a variable grouped by another variable.
It offers options to customize the plot's colors, title, axis labels, and more.
}
\examples{
data("sig_stad", package = "IOBR")
table(sig_stad$Subtype, sig_stad$Lauren)
percent_bar_plot(input = sig_stad, x = "Subtype", y = "Lauren", axis_angle = 60)
}
\author{
Dongqiang Zeng
}
