% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_bar.R
\name{pie_chart}
\alias{pie_chart}
\title{pie_chart}
\usage{
pie_chart(
  input,
  var,
  var2 = NULL,
  type = 2,
  show_freq = FALSE,
  color = NULL,
  palette = "jama",
  title = NULL,
  text_size = 10,
  title_size = 20,
  add_sum = FALSE
)
}
\arguments{
\item{input}{data frame}

\item{var}{target}

\item{type}{plot type}

\item{add_sum}{}
}
\description{
pie_chart
}
