% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModel.R
\name{PlotAUC}
\alias{PlotAUC}
\title{Plot AUC}
\usage{
PlotAUC(train.x, train.y, test.x, test.y, model, foldername, modelname)
}
\arguments{
\item{train.x}{A matrix or data frame containing the training features.}

\item{train.y}{A vector containing the training labels.}

\item{test.x}{A matrix or data frame containing the testing features.}

\item{test.y}{A vector containing the testing labels.}

\item{model}{A fitted model object, typically from glmnet.}

\item{foldername}{A character string specifying the folder name where the plot will be saved.}

\item{modelname}{A character string specifying the name of the model, used in the plot title and filename.}
}
\value{
A ggplot object containing the plotted ROC curves.
}
\description{
This function plots the AUC (Area Under the Curve) ROC (Receiver Operating Characteristic) curves for the given model on both training and testing datasets.
}
\examples{
# Assuming train.x, train.y, test.x, and test.y are already defined
model <- glmnet::cv.glmnet(x = train.x, y = train.y, family = "binomial")
p <- PlotAUC(train.x = train.x, train.y = train.y, test.x = test.x, test.y = test.y,
             model = model, foldername = "plots", modelname = "lasso_model")
print(p)
}
