% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_helper.R
\name{plotPurity}
\alias{plotPurity}
\title{plotPurity}
\usage{
plotPurity(
  scores,
  samples = "all_samples",
  platform = c("affymetrix", "agilent", "illumina"),
  output.dir = "estimated_purity_plots"
)
}
\arguments{
\item{samples}{default is all_samples}

\item{platform}{"affymetrix", "agilent", "illumina"}

\item{output.dir}{default is estimated_purity_plots}
}
\description{
plotPurity
}
