% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{PlotTimeROC}
\alias{PlotTimeROC}
\title{Plot Time ROC}
\usage{
PlotTimeROC(train.x, train.y, test.x, test.y, model, foldername, modelname)
}
\arguments{
\item{modelname}{}
}
\value{

}
\description{
Plot Time ROC
}
