% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{PrognosticModel}
\alias{PrognosticModel}
\title{Title}
\usage{
PrognosticModel(
  x,
  y,
  scale = T,
  seed = "123456",
  train_ratio = 0.7,
  nfold = 10,
  plot = T
)
}
\arguments{
\item{x}{input matrix or data.frame where samples are in rows and features in columns; The first column of x is the sample ID of which column is "ID".}

\item{y}{input matrix or data.frame with three column. Column names are "ID", "time", "status"}

\item{scale}{A logistic: should the x be scaled, default is TRUE.}

\item{seed}{default 123456}

\item{train_ratio}{Value between 0-1; The ratio was used to split the x and y into training and testing data.}

\item{nfold}{nfold default 10}

\item{plot}{}
}
\value{
a list contain the results of 2 model (Lasso, Ridge) and the input train data.
}
\description{
Title
}
\examples{
load("./data/crc_response.rda")
load("./data/tme_combine.rda")
x <- tme_combine
y <- na.omit(response[, c("ID", "OS.time", "OS")])
colnames(y) <- c("ID", "time", "status")
PrognosticResult <- PrognosticModel(x = x, y = y, scale = T, seed = "123456", train_ratio = 0.7)
}
