% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{PrognosticResult}
\alias{PrognosticResult}
\title{Prognostic Result}
\usage{
PrognosticResult(model, train.x, train.y, test.x, test.y)
}
\arguments{
\item{test.y}{}
}
\value{

}
\description{
Prognostic Result
}
