% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{rawEnrichmentAnalysis}
\alias{rawEnrichmentAnalysis}
\title{Calculated raw xCell enrichment scores}
\usage{
rawEnrichmentAnalysis(expr, signatures, genes, file.name = NULL)
}
\arguments{
\item{expr}{the gene expression data set. A matrix with row names as symbols and columns as samples.}

\item{signatures}{a GMT object of signatures.}

\item{genes}{list of genes to use in the analysis.}

\item{file.name}{string for the file name for saving the scores. Default is NULL.}
}
\value{
the raw xCell scores
}
\description{
rawEnrichmentAnalysis Returns the raw xCell cell types enrichment scores.
}
