% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_iobr.R
\name{rbind_iobr}
\alias{rbind_iobr}
\title{Row Bind Multiple Data Sets}
\usage{
rbind_iobr(data1, data2, data3 = NULL)
}
\arguments{
\item{data1}{Data frame or matrix. First dataset to combine.}

\item{data2}{Data frame or matrix. Second dataset to combine.}

\item{data3}{Data frame or matrix, optional. Additional dataset to combine. Default is NULL.}
}
\value{
Combined data frame or matrix resulting from row binding the input datasets.
}
\description{
Combines two or three data frames or matrices vertically using the `rbind` operation. Ensures compatibility of input data before binding.
}
\examples{
data1 <- data.frame(A = 1:5, B = letters[1:5])
data2 <- data.frame(A = 6:10, B = letters[6:10])
combined_data <- rbind_iobr(data1, data2)
# With three datasets
data3 <- data.frame(A = 11:15, B = letters[11:15])
combined_data <- rbind_iobr(data1, data2, data3)
}
