% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModel.R
\name{RegressionResult}
\alias{RegressionResult}
\title{Regression Result}
\usage{
RegressionResult(train.x, train.y, test.x, test.y, model)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Regression Result
}
