% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_batcheffect.R
\name{remove_batcheffect}
\alias{remove_batcheffect}
\title{remove batch effect of two or three expression set}
\usage{
remove_batcheffect(
  eset1,
  eset2,
  eset3 = NULL,
  log2 = TRUE,
  check_eset = TRUE,
  labels = NULL,
  palette = "nrc",
  path = NULL,
  save_plot = FALSE
)
}
\arguments{
\item{eset1}{expression set-1, normalized data}

\item{eset2}{expression set-2, normalized data}

\item{eset3}{expression set-3, default is null}

\item{log2}{default is TRUE}

\item{check_eset}{default is true}

\item{labels}{labels for expression sets}

\item{palette}{palette from IOBR::palettes, default is nature}

\item{path}{default is NULL}

\item{save_plot}{default is FALSE, if true, path could be set to folder name}
}
\description{
remove batch effect of two or three expression set
}
