% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_genes.R
\name{remove_duplicate_genes}
\alias{remove_duplicate_genes}
\title{Remove duplicate gene symbol on a gene expression data}
\usage{
remove_duplicate_genes(eset, column_of_symbol, method = "mean")
}
\arguments{
\item{eset}{Gene expression data with gene symbol in `column_of_symbol`}

\item{column_of_symbol}{name of column contained gene symbols}

\item{method}{method used to filter duplicate genes; default is mean value}
}
\value{
A modified gene expression data frame with duplicates removed and selected method applied for aggregation.
}
\description{
This function removes duplicate gene symbols from a gene expression dataset. It offers the option to aggregate duplicate gene expressions using different methods such as mean or standard deviation.
}
\examples{

# loading eset
data("eset_stad", package = "IOBR")
# annotation
eset_stad <- anno_eset(eset = eset_stad, annotation = anno_rnaseq)
eset_stad <- rownames_to_column(eset_stad, var = "id")

# Creating duplicate gene names
eset_stad[2:3, "id"] <- "MT-CO1"
# Counting the number of identical names
summary(duplicated(eset_stad$id))
# De-duplication of rows with the same gene name using the average value
eset_stad<-remove_duplicate_genes(eset = eset_stad, column_of_symbol = "id", method = "mean")
summary(duplicated(eset_stad$id))
}
\author{
Dongqiang Zeng
}
