% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_names.R
\name{remove_names}
\alias{remove_names}
\title{Batch to transform patterns to special character}
\usage{
remove_names(
  input_df,
  variable = "colnames",
  patterns_to_na = patterns_to_na,
  patterns_space = NULL
)
}
\arguments{
\item{input_df}{input data frame}

\item{patterns_to_na}{patterns that will be transform into NA}

\item{patterns_space}{patterns that will be transform into space}

\item{variables}{colunm names or names of variable}
}
\value{

}
\description{
Batch to transform patterns to special character
}
