% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_names.R
\name{remove_names}
\alias{remove_names}
\title{Remove Patterns from Column Names or Variables}
\usage{
remove_names(
  input_df,
  variable = "colnames",
  patterns_to_na = patterns_to_na,
  patterns_space = NULL
)
}
\arguments{
\item{input_df}{Input data frame.}

\item{variable}{Column to modify: "colnames" for column names, or a specific column name. Default is "colnames".}

\item{patterns_to_na}{Vector of patterns to replace with empty string. Default uses predefined patterns.}

\item{patterns_space}{Vector of patterns to replace with spaces. Default is NULL.}
}
\value{
Modified data frame with patterns replaced.
}
\description{
This function modifies column names or specified variables in a data frame by replacing
specified patterns with NA or spaces.
}
\examples{
data("imvigor210_sig", package = "IOBR")
input <- remove_names(imvigor210_sig, variable = "colnames", patterns_to_na = patterns_to_na, patterns_space = NULL)
}
