% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{RemoveBatchEffect}
\alias{RemoveBatchEffect}
\title{Remove batch effect of expression set}
\usage{
RemoveBatchEffect(cancer.exp, immune.exp, immune.cellType)
}
\arguments{
\item{immune.cellType}{}
}
\description{
Remove batch effect of expression set
}
