% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_time.R
\name{roc_time}
\alias{roc_time}
\title{Title}
\usage{
roc_time(
  input,
  vars,
  time = "time",
  status = "status",
  time_point = 12,
  time_type = "month",
  palette = "jama",
  cols = "normal",
  seed = 1234,
  show_col = FALSE,
  path = NULL,
  main = "PFS",
  index = 1,
  fig.type = "pdf",
  width = 5,
  height = 5.2
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
