% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_time.R
\name{roc_time}
\alias{roc_time}
\title{Time-dependent ROC Curve for Survival Analysis}
\usage{
roc_time(
  input,
  vars,
  time = "time",
  status = "status",
  time_point = 12,
  time_type = "month",
  palette = "jama",
  cols = "normal",
  seed = 1234,
  show_col = FALSE,
  path = NULL,
  main = "PFS",
  index = 1,
  fig.type = "pdf",
  width = 5,
  height = 5.2
)
}
\arguments{
\item{input}{Data frame. Input data containing variables for analysis.}

\item{vars}{Character vector. Variable(s) to be evaluated.}

\item{time}{Character. Name of the time variable in the input data. Default is "time".}

\item{status}{Character. Name of the status variable in the input data. Default is "status".}

\item{time_point}{Integer or vector. Time point(s) for ROC analysis. Default is 12.}

\item{time_type}{Character. Time unit (e.g., "month" or "day"). Default is "month".}

\item{palette}{Character. Color palette for the plot. Default is "jama".}

\item{cols}{Character. Color scheme for the plot. Default is "normal".}

\item{seed}{Integer. Random seed for reproducibility. Default is 1234.}

\item{show_col}{Logical. Whether to display the color palette. Default is FALSE.}

\item{path}{Character or NULL. Path to save the plot. Default is NULL.}

\item{main}{Character. Main title of the plot. Default is "PFS".}

\item{index}{Integer. Index for plot filename. Default is 1.}

\item{fig.type}{Character. Output file type (e.g., "pdf", "png"). Default is "pdf".}

\item{width}{Numeric. Width of the plot. Default is 5.}

\item{height}{Numeric. Height of the plot. Default is 5.2.}
}
\value{
A ggplot object representing the time-dependent ROC plot.
}
\description{
Generates a time-dependent Receiver Operating Characteristic (ROC) plot to evaluate the predictive performance of one or more variables in survival analysis. Calculates the Area Under the Curve (AUC) for each specified time point and variable, and creates a multi-line ROC plot with annotated AUC values.
}
\examples{
data("tcga_stad_sig", package = "IOBR")
data("pdata_stad", package = "IOBR")
input <- merge(pdata_stad, tcga_stad_sig, by = "ID")
roc_time(
  input = input, vars = c("Pan_F_TBRs", "CD_8_T_effector", "Immune_Checkpoint"),
  time = "time", status = "OS_status", time_point = 12, path = "result", main = "OS"
)
}
