% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_matrix.R
\name{scale_matrix}
\alias{scale_matrix}
\title{Scale matrix}
\usage{
scale_matrix(matrix, log2matrix = TRUE, manipulate = FALSE)
}
\arguments{
\item{matrix}{matrix}

\item{log2matrix}{default is TRUE}

\item{manipulate}{}
}
\description{
Scale matrix
}
