% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_matrix.R
\name{scale_matrix}
\alias{scale_matrix}
\title{Scale and Manipulate a Matrix}
\usage{
scale_matrix(matrix, log2matrix = TRUE, manipulate = TRUE)
}
\arguments{
\item{matrix}{The matrix to be scaled and manipulated.}

\item{log2matrix}{Logical indicating whether to apply log2 transformation. Default is TRUE.}

\item{manipulate}{Logical indicating whether to perform feature manipulation. Default is TRUE.}
}
\value{
A data frame of the scaled and manipulated matrix.
}
\description{
This function scales a given matrix, optionally applies logarithmic transformation,
and performs feature manipulation to remove problematic variables.
}
\examples{
data("eset_gse62254", package = "IOBR")
eset2 <- scale_matrix(eset_gse62254, log2matrix = FALSE, manipulate = TRUE)
}
\author{
Dongqiang Zeng
}
