% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_method.R
\name{select_method}
\alias{select_method}
\title{Choose which method to conduct following analysis}
\usage{
select_method(data, method = "ssGSEA")
}
\arguments{
\item{data}{signature matrix}

\item{method}{choose method to analysis, contain three methodologies : PCA, ssGSEA, zScore}
}
\value{
filtered signature score matrix
}
\description{
Choose which method to conduct following analysis
}
\examples{
data(eset_stad, package = "IOBR")
eset <- count2tpm(eset_stad, idType = "ensembl")
head(eset)
eset <- scale_matrix(eset, manipulate = TRUE)

res <- calculate_sig_score(eset = eset, signature = signature_collection[1:4], method = "integration")
head(res)
res <- select_method(res, method = "PCA")
head(res)
}
