% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_method.R
\name{select_method}
\alias{select_method}
\title{Choose which method to conduct analysis}
\usage{
select_method(score = score, choosemethod = "ssGSEA")
}
\arguments{
\item{score}{TME signature score matrix}

\item{choosemethod}{choosemethod to analysis, contain three methodologies : PCA, ssGSEA, zScore}
}
\value{
filterd TME signature score matrix
}
\description{
Choose which method to conduct analysis
}
