% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_method.R
\name{select_method}
\alias{select_method}
\title{Choose which method to conduct following analysis}
\usage{
select_method(data, method = "ssGSEA")
}
\arguments{
\item{data}{signature matrix}

\item{method}{choose method to analysis, contain three methodologies : PCA, ssGSEA, zScore}
}
\value{
filtered signature score matrix
}
\description{
Choose which method to conduct following analysis
}
