% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_box.R
\name{sig_box}
\alias{sig_box}
\title{signature box plot}
\usage{
sig_box(
  data,
  signature,
  variable,
  angle_x_text = 0,
  hjust = 0,
  palette = "nrc",
  cols = NULL,
  jitter = FALSE,
  point_size = 5,
  show_pvalue = TRUE,
  return_stat_res = FALSE
)
}
\arguments{
\item{data}{input data}

\item{signature}{name of signature}

\item{variable}{name of category variable}

\item{angle_x_text}{angle_x_text}

\item{hjust}{hjust}

\item{palette}{default is nature palette}

\item{cols}{default is null}

\item{jitter}{default is FALSE}

\item{point_size}{size of point}

\item{show_pvalue}{default is true}

\item{return_stat_res}{default is FALSE}
}
\value{

}
\description{
signature box plot
}
\examples{
sig_box(data = tcga_stad_pdata, signature = "TMEscore_plus", variable = "subtype",jitter = T)
}
