% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_forest.R
\name{sig_forest}
\alias{sig_forest}
\title{Forest plot of survival analysis results.}
\usage{
sig_forest(
  data,
  signature,
  pvalue = "P",
  HR = "HR",
  CI_low_0.95 = "CI_low_0.95",
  CI_up_0.95 = "CI_up_0.95",
  n = 10,
  max_character = 25,
  discrete_width = 35,
  color_option = 1
)
}
\arguments{
\item{data}{survival result}

\item{signature}{column of signatures}

\item{pvalue}{P value of survival result}

\item{HR}{Hazard ratio of survival result}

\item{CI_low_0.95}{confidence interval}

\item{CI_up_0.95}{confidence interval}

\item{n}{maximum of signatures}

\item{max_character}{maximum of character}

\item{discrete_width}{discrete width, default is 35}

\item{color_option}{default is 1, other option:2, 3}
}
\value{
A ggplot object representing the forest plot.
}
\description{
This function, `sig_forest`, generates a forest plot displaying the hazard ratios (HRs), confidence intervals, and P-values for survival analysis signatures. It is designed to handle dataframes containing survival analysis results and allows customization of the plot features and aesthetics.
}
\examples{
sig_surv_result<- batch_surv(pdata = pdata_sig_tme_binary,variable = c(100:ncol(pdata_sig_tme_binary)))
sig_forest(data = sig_surv_result, signature = "ID")

}
\author{
Dongqiang Zeng
}
