% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_heatmap.R
\name{sig_heatmap}
\alias{sig_heatmap}
\title{simple heatmap}
\usage{
sig_heatmap(
  input,
  ID = "ID",
  features,
  group,
  palette = 2,
  palette_group = "jama",
  show_col = F,
  show_palettes = F,
  cols_group = NULL,
  show_plot = T,
  width = 8,
  height = NULL,
  size_col = 10,
  size_row = 8,
  angle_col = 90,
  column_title = NULL,
  row_title = NULL,
  show_heatmap_col_name = F,
  path = NULL,
  index = NULL
)
}
\arguments{
\item{input}{data frame with `ID`, `group` and features}

\item{ID}{default is ID}

\item{features}{features}

\item{group}{category}

\item{palette}{default is 2, options: 1,2,3}

\item{palette_group}{default is jama}

\item{show_col}{show color}

\item{show_palettes}{show palettes}

\item{cols_group}{user can define the colors of group}

\item{show_plot}{show plot}

\item{width}{default is 8}

\item{height}{height of figure}

\item{size_col}{font size of column name}

\item{size_row}{font size of row name}

\item{angle_col}{angle of column name}

\item{column_title}{title of column}

\item{row_title}{title of row}

\item{show_heatmap_col_name}{show heat map column name}

\item{path}{save path, folder name}

\item{index}{default is null}
}
\description{
simple heatmap
}
