% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_roc.R
\name{sig_roc}
\alias{sig_roc}
\title{Drawing multiple AUC into one graph}
\usage{
sig_roc(
  data,
  response,
  variables,
  fig.path = ".",
  main = NULL,
  file.name = NULL,
  palette = "jama",
  cols = NULL,
  alpha = 1,
  compare = FALSE,
  smooth = TRUE,
  compare_method = "bootstrap",
  boot.n = 100
)
}
\arguments{
\item{data}{A data frame containing the response variable and predictors.}

\item{response}{The name of the response variable in the data frame, typically a binary outcome.}

\item{variables}{A vector of column names from `data` representing predictors for which ROC curves will be plotted.}

\item{fig.path}{Optional; the path where the output plot will be saved, defaults to the current working directory.}

\item{main}{The main title of the plot.}

\item{file.name}{Optional; the base name for the output file, defaults to "0-ROC of multiple variables" if not specified.}

\item{palette}{default is `jama`, if number of variables is bigger than colors of palettes, `random` color will be applied}

\item{cols}{Optional; a vector of colors to use for the curves. If not provided, colors are taken from the specified palette.}

\item{alpha}{Transparency level of the plot colors; default is 1 (opaque).}

\item{compare}{Logical; if TRUE, compares the ROC curves using statistical tests; default is FALSE.}

\item{smooth}{Logical; if TRUE, the ROC curves are smoothed; default is TRUE}

\item{compare_method}{The method used for comparison if `compare` is TRUE; defaults to "bootstrap". Other options include "delong" and "venkatraman".}

\item{boot.n}{Number of bootstrap replicates if the bootstrap method is used; default is 100.}
}
\value{
A list containing the AUC values, legend names for the plot, and optionally the comparison results.
}
\description{
This function `sig_roc` creates ROC curves for evaluating the performance of one or more predictors in a binary classification task. It is primarily designed to assess how well variables can distinguish between two classes, such as disease vs. no disease. The function generates a plot for each variable and optionally compares their Areas Under the Curve (AUCs) using statistical tests.
}
\examples{
data("tcga_stad_pdata", package = "IOBR")
sig_roc(data = tcga_stad_pdata, response = "OS_status", variables = c("TMEscore_plus", "GZMB", "GNLY"))
}
\author{
Dongqiang Zeng
}
