% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_roc.R
\name{sig_roc}
\alias{sig_roc}
\title{Drawing multiple AUC into one graph}
\usage{
sig_roc(
  data,
  response,
  variables,
  fig.path = ".",
  main = NULL,
  file.name = NULL,
  palette = "jama",
  cols = NULL,
  alpha = 1,
  compare = FALSE,
  compare_method = "bootstrap",
  boot.n = 100
)
}
\arguments{
\item{data}{data frame with response and variables}

\item{response}{response name of data}

\item{variables}{variables}

\item{fig.path}{default is current working directory}

\item{main}{main title of plot}

\item{file.name}{file.name, default is}

\item{palette}{default is `jama`, if number of variables is bigger than colors of palettes, `random` color will be applied}

\item{cols}{users can provide cols manually}

\item{alpha}{default is 1}

\item{compare}{default is FALSE}

\item{compare_method}{default is bootstrap, other option: “delong”, “venkatraman”}

\item{boot.n}{default is 100 when bootstrap is chosen.}
}
\description{
Drawing multiple AUC into one graph
}
\author{
Dongqiang Zeng
}
