% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score.R
\docType{data}
\name{signature_score_calculation_methods}
\alias{signature_score_calculation_methods}
\title{List of supported signature score calculation methods}
\format{An object of class \code{character} of length 4.}
\usage{
signature_score_calculation_methods
}
\description{
The methods currently supported are
`PCA`, `ssGSEA`, `z-score`, `Integration`
}
\details{
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
