% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigScore.R
\name{sigScore}
\alias{sigScore}
\title{Calculate siganture score using PCA function}
\usage{
sigScore(eset, methods = "PCA")
}
\arguments{
\item{eset}{normalized count matrix; rows are all genes in the signature that shall be summarized into one score; columns are samples}

\item{methods}{character vector defining whether signature scores shall be based on principal component 1 ("PC", default) or z-scores (other value)}
}
\value{
numeric vector of length ncol(eset); a score summarizing the rows of eset
}
\description{
Calculate siganture score using PCA function
}
\author{
Dorothee Nickles, Dongqiang Zeng
}
