% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModel.R
\name{SplitTrainTest}
\alias{SplitTrainTest}
\title{Split Data into Training and Testing Sets}
\usage{
SplitTrainTest(x, y, train_ratio, type = c("binomial", "survival"), seed)
}
\arguments{
\item{x}{Predictor matrix or data frame.}

\item{y}{Outcome vector or matrix.}

\item{train_ratio}{Proportion for training (0-1).}

\item{type}{`"binomial"` or `"survival"`.}

\item{seed}{Random seed.}
}
\value{
List with train.x, train.y, test.x, test.y, train_sample.
}
\description{
Divides dataset into training and testing sets.
}
\examples{
\donttest{
data_matrix <- matrix(rnorm(200), ncol = 2)
outcome_vector <- rbinom(100, 1, 0.5)
split_data <- SplitTrainTest(data_matrix, outcome_vector, train_ratio = 0.7, type = "binomial", seed = 123)
}
}
