% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModel.R
\name{SplitTrainTest}
\alias{SplitTrainTest}
\title{Split Train and Test data}
\usage{
SplitTrainTest(x, y, train_ratio, type, seed)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Split Train and Test data
}
