% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_survival.R
\name{subgroup_survival}
\alias{subgroup_survival}
\title{Title Extract hazard ratio and confidence intervals from a coxph object of subgroup analysis}
\usage{
subgroup_survival(pdata, time_name = "time", status_name = "status",
  variables, object)
}
\arguments{
\item{pdata}{data combine variables, follow up time, and outcome;}

\item{time_name}{column name of follow up time}

\item{status_name}{column name of outcome event}

\item{variables}{object that have several levles;}

\item{object}{target that use to calculate coxph fit;
If target has two levels which name `High` and `Low` group, please transform `High` into `0` and `Low` into `1` before executing the command}
}
\description{
Title Extract hazard ratio and confidence intervals from a coxph object of subgroup analysis
}
\examples{

#source data and filter NA
data(subgroup_data)
input <- subgroup_data \%>\% filter(time > 0) \%>\% filter(!is.na(status)) \%>\% filter(!is.na(AJCC_stage))

##for binary variable
data1 <- subgroup_survival(pdata = input,time_name ="time", status_name = "status",
variables = c("ProjectID", "AJCC_stage"), object ="score_binary" )
data1

##for continue variables
data2 <- subgroup_survival(pdata = input, time_name = "time", status_name = "status",
variables = c("ProjectID", "AJCC_stage"), object = "score" )
data2
}
\author{
Dongqiang Zeng
}
