% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_group.R
\name{surv_group}
\alias{surv_group}
\title{Drawing KM-plot to a Categorical variable with two or more groups}
\usage{
surv_group(
  input_pdata,
  target_group,
  ID = "ID",
  levels = c("High", "Low"),
  reference_group = "High",
  project = NULL,
  time = "time",
  status = "status",
  time_type = "month",
  break_month = "auto",
  cols = NULL,
  palette = "jama",
  mini_sig = "score",
  save_path = paste0("KMplot"),
  fig.type = "pdf",
  index = 1,
  width = 6,
  height = 6.5,
  font.size.table = 3
)
}
\arguments{
\item{input_pdata}{DATA}

\item{target_group}{target of binary variable}

\item{ID}{identifier of data}

\item{levels}{level of variable}

\item{reference_group}{default is}

\item{project}{name of main title}

\item{time}{column name of follow up time}

\item{status}{status}

\item{time_type}{month or day}

\item{break_month}{break of time}

\item{cols}{if null, palette should be set}

\item{palette}{default is `jama`, if cols is null}

\item{mini_sig}{prefix of variable}

\item{save_path}{default is KMplot}

\item{fig.type}{default is pdf, other option is png}

\item{index}{folder name}

\item{width}{default is 6}

\item{height}{default is 6.5}

\item{font.size.table}{default is 62}
}
\description{
Drawing KM-plot to a Categorical variable with two or more groups
}
\examples{

define break time
}
\author{
Dongqiang Zeng
}
