% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_group.R
\name{surv_group}
\alias{surv_group}
\title{Drawing KM-plot to a Categorical variable with two or more groups}
\usage{
surv_group(
  input_pdata,
  target_group,
  ID = "ID",
  levels = c("High", "Low"),
  reference_group = "High",
  project = NULL,
  time = "time",
  status = "status",
  time_type = "month",
  break_month = "auto",
  cols = NULL,
  palette = "jama",
  mini_sig = "score",
  save_path = paste0("KMplot"),
  fig.type = "pdf",
  index = 1,
  width = 6,
  height = 6.5,
  font.size.table = 3
)
}
\arguments{
\item{input_pdata}{Data frame containing the survival time, status, and group variables.}

\item{target_group}{The name of the categorical variable in `input_pdata` used to define groups for survival analysis.}

\item{ID}{Identifier for the data rows; default is "ID".}

\item{levels}{A vector of the levels or groups within the target group variable; defaults to c("High", "Low").}

\item{reference_group}{The reference group against which other groups are compared in survival analysis; defaults to the first level if not specified.}

\item{project}{Name for the main title of the plot and part of the filename for saving the plot.}

\item{time}{The column name in `input_pdata` representing follow-up time; default is "time".}

\item{status}{The column name in `input_pdata` representing the event status; default is "status".}

\item{time_type}{Unit of follow-up time, either "month" or "day"; default is "month".}

\item{break_month}{Time intervals for breaking the follow-up period; default is "auto" which calculates based on data.}

\item{cols}{Color vector for plots; if NULL, colors are selected based on `palette`.}

\item{palette}{Color palette for plotting; default is "jama".}

\item{mini_sig}{A prefix used for variable naming in plots and files; default is "score".}

\item{save_path}{Directory where the output files will be stored; default is "KMplot".}

\item{fig.type}{The format for saving plots, either "pdf" or "png"; default is "pdf".}

\item{index}{Folder name or prefix for saving the plot; used in file naming.}

\item{width}{Width of the output plot; default is 6 inches.}

\item{height}{Height of the output plot; default is 6.5 inches.}

\item{font.size.table}{Font size for tables included in the plot; default is 62.}
}
\value{
A list containing the ggsurvplot object and optionally comparison results if comparisons are enabled.
}
\description{
This function produces Kaplan-Meier survival plots for a given categorical variable with two or more groups. It allows customization of the analysis and visualization, including defining the time scale (months or days), breaking the follow-up time into intervals, selecting color palettes, and saving the plots. The function supports comparisons between groups using survival statistics and can handle additional grouping variables.
}
\examples{
data("tcga_stad_pdata", package = "IOBR")
surv_group(input_pdata = tcga_stad_pdata, target_group = "TMEscore_plus_binary", time = "time", status = "OS_status")

}
\author{
Dongqiang Zeng
}
