% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_rna_pres.R
\name{tcga_rna_preps}
\alias{tcga_rna_preps}
\title{Title tcga_rna_preps}
\usage{
tcga_rna_preps(
  eset,
  id_type = c("ensembl", "symbol"),
  input_type = c("log2count", "count"),
  output = c("tumor", "tumor_normal"),
  output_type = c("tpm", "log2tpm", "count"),
  annotation = TRUE
)
}
\arguments{
\item{eset}{A matrix or data frame representing RNA-seq gene expression data, typically from The Cancer Genome Atlas (TCGA).}

\item{id_type}{(optional, default = "ensembl"): Specifies the type of gene identifier to be used, accepting values "ensembl" or "symbol".}

\item{input_type}{(optional, default = "log2count"): Indicates the input data type, accepting values "log2count" or "count".}

\item{output}{(optional, default = "tumor"): Specifies the output sample type, accepting values "tumor" or "tumor_normal".}

\item{output_type}{(optional, default = "tpm"): Indicates the output data type, accepting values "tpm", "log2tpm", or "count".}

\item{annotation}{(optional, default = TRUE): A logical value determining whether gene annotation should be applied.}
}
\description{
The tcga_rna_preps function is designed to preprocess TCGA RNA-seq data for downstream analysis. Upon executing the function, it performs various operations based on the specified parameters, including modifying sample types, transforming input data, and annotating genes. The function returns the preprocessed RNA-seq gene expression data in the specified output format.
}
\examples{
# Expression matrix downloaded from UCSC Xena datasets: https://xenabrowser.net/datapages/
eset <- tcga_rna_preps(eset = eset,  id_type = "ensembl", input_type = "log2count", output = "tumor", output_type = "count", annotation = TRUE)

}
\author{
Dongqiang Zeng
}
