% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\docType{data}
\name{timer_available_cancers}
\alias{timer_available_cancers}
\title{TIMER Available Cancer Types}
\format{
An object of class \code{character} of length 32.
}
\usage{
timer_available_cancers
}
\description{
Character vector of cancer types supported by TIMER deconvolution.
TIMER signatures are cancer-specific.
}
\keyword{datasets}
