% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tme_cluster.R
\name{tme_cluster}
\alias{tme_cluster}
\title{Identification of TME cluster}
\usage{
tme_cluster(
  input,
  features = NULL,
  pattern = NULL,
  id = NULL,
  scale = TRUE,
  method = "kmeans",
  min_nc = 2,
  max.nc = 6
)
}
\arguments{
\item{input}{A data frame containing the input dataset for TME clustering analysis.}

\item{features}{A vector specifying the features (variables) to be used for clustering. Default is NULL, which uses all columns as features.}

\item{pattern}{A regular expression pattern for selecting features based on column names. Default is NULL.}

\item{id}{A character string specifying the column in the input data frame to be used as identifiers. Default is NULL, which uses row names as identifiers.}

\item{scale}{A logical value indicating whether to scale the selected features. Default is TRUE.}

\item{method}{A character string specifying the clustering method to use. Default is "kmeans".}

\item{min_nc}{An integer specifying the minimum number of clusters to evaluate. Default is 2.}

\item{max.nc}{An integer specifying the maximum number of clusters to evaluate. Default is 6.}
}
\value{
A data frame with clusters appended as a new column.
}
\description{
This function is designed to perform TME (Tumor Microenvironment) clustering analysis using the specified input dataset. The function allows for customization of various parameters to conduct clustering analysis, including the selection of features, scaling options, clustering method, and the range of cluster numbers to evaluate.
}
\examples{
data("tcga_stad_sig", package = "IOBR")
res <- tme_cluster(input = tcga_stad_sig, features = NULL, pattern = "xCell", id = "ID", method = "kmeans", min_nc = 2, max.nc = 6)
sig_heatmap(input = res, features = colnames(res)[3:ncol(res)], group = "cluster")
}
\author{
Dongqiang Zeng
}
