% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo_tme.R
\docType{data}
\name{tme_deconvolution_methods}
\alias{tme_deconvolution_methods}
\title{Collection of tumor microenvironment cell fraction deconvolution methods.}
\format{
An object of class \code{character} of length 11.
}
\usage{
tme_deconvolution_methods
}
\description{
The methods currently supported are
`mcpcounter`, `epic`, `xcell`, `cibersort`, `cibersort_abs`, `ips`, `estimate`, `svr`,`lsei`,`timer`,`quantiseq`
}
\details{
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
