% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform features with Inf or NA to other number}
\usage{
transform_data(data = data, feature = feature, data_type, into = 0)
}
\arguments{
\item{data}{data with Inf}

\item{feature}{variables}

\item{data_type}{is data NA or infinite}

\item{into}{transform data into 0 or 1, or others, default is 0}
}
\value{
minipulated data
}
\description{
Transform features with Inf or NA to other number
}
\examples{

}
\author{
Dongqiang Zeng
}
