% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{xCellAnalysis}
\alias{xCellAnalysis}
\title{The xCell analysis pipeline}
\usage{
xCellAnalysis(
  expr,
  signatures = NULL,
  genes = NULL,
  spill = NULL,
  rnaseq = TRUE,
  file.name = NULL,
  scale = TRUE,
  alpha = 0.5,
  save.raw = FALSE,
  cell.types.use = NULL
)
}
\arguments{
\item{expr}{the gene expression data set. A matrix with row names as symbols and columns as samples.}

\item{signatures}{a GMT object of signatures.}

\item{genes}{list of genes to use in the analysis.}

\item{spill}{the Spillover object for adjusting the scores.}

\item{rnaseq}{if true than use RNAseq spillover and calibration paramters, else use array parameters.}

\item{file.name}{string for the file name for saving the scores. Default is NULL.}

\item{scale}{if TRUE, uses scaling to trnasform scores using fit.vals}

\item{alpha}{a value to override the spillover alpha parameter. Deafult = 0.5}

\item{save.raw}{TRUE to save a raw}

\item{cell.types.use}{a character list of the cell types to use in the analysis. If NULL runs xCell with all cell types.
The spillover compensation step may over compensate, thus it is always better to run xCell with a list of cell types that are expected
to be in the mixture. The names of cell types in this list must be a subset of the cell types that are inferred by xCell.}
}
\value{
the adjusted xCell scores
}
\description{
The xCell analysis pipeline
}
