% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{xCellSignifcanceRandomMatrix}
\alias{xCellSignifcanceRandomMatrix}
\title{Calculate significance p-values for the null hypothesis that the cell type is not present in the mixture using a random matrix.}
\usage{
xCellSignifcanceRandomMatrix(
  scores,
  expr,
  spill,
  alpha = 0.5,
  nperm = 250,
  file.name = NULL
)
}
\arguments{
\item{scores}{the xCell scores.}

\item{expr}{the input expression matrix.}

\item{spill}{the Spillover object for adjusting the scores.}

\item{alpha}{a value to override the spillover alpha parameter. Deafult = 0.5}

\item{nperm}{number of samples in the shuffled matrix, default = 250}

\item{file.name}{file name to write the p-values table.}
}
\value{
a list with the p-values, the xcell scores of the shuffled data and the shuffled expression matrix.
}
\description{
xCellSignifcanceRandomMatrix Returns the FDR adjusted p-values of the chance that the cell is not present in the mixture.
}
